% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{addMedianErrorBarValues}
\alias{addMedianErrorBarValues}
\title{Add to data MAD min/max y-values for plotting median error bars}
\usage{
addMedianErrorBarValues(dat)
}
\arguments{
\item{dat}{tibble including at least columns for Sample,CellType & Density;
Also must include rows with SPOT=="Median" representing median of
all spots for each Sample+CellType}
}
\value{
original tibbles with added columns for TotalDensity, mad, ymin, ymax
}
\description{
For each SPOT=="Median" row in tibble, add MAD of densities for
all SPOTs for a specific Sample+CellType. Then add ymin and ymax values
representing total median density +/- MAD
}
