% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_meta.R
\name{assignCellType}
\alias{assignCellType}
\title{Based on MarkerCombinationAnnotation table from flattenMetaData(), assign a cell type
to a combination in marker combination table}
\usage{
assignCellType(combinationRow, cellTypes, conditionalTypes, totalCellCount)
}
\arguments{
\item{combinationRow}{table row from marker combination table}

\item{cellTypes}{table describing known marker combinations, containing at least
Marker_combination and Cell_type columns}

\item{conditionalTypes}{table of conditional cell types (*NOTE* this is not implemented yet)}
}
\value{
character string indicating cell type of given combination
}
\description{
Based on MarkerCombinationAnnotation table from flattenMetaData(), assign a cell type
to a combination in marker combination table
}
