% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{calcDistancesFromTumorBoundary}
\alias{calcDistancesFromTumorBoundary}
\title{Calculate distances from each cell to nearest point on tumor boundary}
\usage{
calcDistancesFromTumorBoundary(dat, aFiles = NULL, haloAnnotations = NULL,
  writeCSVfiles = TRUE, outDir = NULL, interfaceBins = (-36:36) * 10)
}
\arguments{
\item{dat}{Halo data for one sample or more samples, loaded from *.rda file(s)}

\item{aFiles}{a vector of Halo boundary annotation files, with 
'.annotations' extension in XML format}

\item{writeCSVfiles}{logical indicating whether to write density and area tables to file; 
default=TRUE}

\item{outDir}{if writeCSVfiles=T, directory to which these files will be written}

\item{interfaceBins}{a vector of distances from tumor interface in which cells will be binned; 
default=(-36:36)*10}
}
\value{
table of all cells including distance to nearest tumor boundary and band assignments
}
\description{
Calculate distances from each cell to nearest point on tumor boundary
}
