% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{calculateAreaTotalFOV}
\alias{calculateAreaTotalFOV}
\title{Calculate total area in each FOV}
\usage{
calculateAreaTotalFOV(dat, metaFiles, areaDir, haloAnnotations = NULL,
  annotationsDirs = NULL, writeCSVfiles = TRUE, maxG = 5)
}
\value{
table of area values for all FOVs
}
\description{
Calculate total area in each FOV
}
