% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{calculateDensity}
\alias{calculateDensity}
\title{Calculate density of set of markers (marker combinations)}
\usage{
calculateDensity(dat, areas, bandAssignments, markerSet, pad,
  funcMarker = NULL, sortByMarker = NULL, writeCSVfiles = TRUE,
  outDir = NULL, statsByBand = FALSE)
}
\arguments{
\item{dat}{tibble containing marker counts}

\item{areas}{tibble containing area by band}

\item{bandAssignments}{tibble containing band assignments for each cell}

\item{markerSet}{vector of markers for which density should be calculated}

\item{pad}{number in microns, amount to trim from border of FOV}

\item{funcMarker}{functional marker for which to calculate density; default=NULL}

\item{sortByMarker}{sort data by density of this marker; default=NULL}

\item{writeCSVfiles}{logical; write density to csv files; default=FALSE}

\item{outDir}{logical; if writeCSVfiles is TRUE, write them to this directory}

\item{statsByBand}{logical; calculate density by interval bands}
}
\value{
tibble containing densities for each marker type
}
\description{
Based on pre-computed areas and band assignments (may be one band, depending on
how area was calculated), calculate densities for a set of markers in each band
}
