% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{calculateFOVStats}
\alias{calculateFOVStats}
\title{Calculate total area and density of entire FOVs}
\usage{
calculateFOVStats(aFiles, dataFile, cellTypeFile, cellTypeName, fovBB, pad,
  maxG, writeCSVfiles = writeCSVfiles, funcMarker = funcMarker, outDir)
}
\arguments{
\item{aFiles}{a vector of Halo boundary annotation files, with 
'.annotations' extension in XML format}

\item{dataFile}{a *.rda file with Halo data for one sample}

\item{cellTypeFile}{file containing cell types to be calculated; each line
is one cell type and each cell type can be either a single
marker name or a comma-separated list of markers}

\item{cellTypeName}{a name to represent the cell types in marker file}

\item{fovBB}{a list with four elements: X0, X1, Y0, Y1, representing the minimum
and maximum coordinates of the FOV to be plotted}

\item{pad}{amount to trim from FOV before calculating}

\item{maxG}{the maximum factor of 10 to use for generating random points for area calculation}

\item{writeCSVfiles}{logical indicating whether to write density and area tables to file;}

\item{funcMarker}{this marker will be added to all markers in cellTypeFile in order to
compare cells that are +/- for this functional marker [ TO DO: REWORD THIS ]}

\item{outDir}{if writeCSVfiles=T, directory to which these files will be written}
}
\value{
a list containing two tables: 'density' and 'area'
}
\description{
Calculate total area and density of FOVs that do not contain 
tumors
}
