% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{calculateInterfaceArea}
\alias{calculateInterfaceArea}
\title{Calculate total area and density of tumor interface}
\usage{
calculateInterfaceArea(dat, aFiles = NULL, haloAnnotations = NULL,
  writeCSVfiles = TRUE, maxG = 5, outDir = NULL,
  interfaceBins = (-36:36) * 10)
}
\arguments{
\item{aFiles}{a vector of Halo boundary annotation files, with 
'.annotations' extension in XML format}

\item{writeCSVfiles}{logical indicating whether to write density and area tables to file; 
default=TRUE}

\item{maxG}{the maximum factor of 10 to use for generating random points for area calculation}

\item{outDir}{if writeCSVfiles=T, directory to which these files will be written}

\item{interfaceBins}{a vector of distances from tumor interface in which cells will be binned; 
default=(-20:20)*10}

\item{data}{Halo data for one sample, loaded from *.rda file}

\item{statsByBand}{logical; return stats broken down by interface bands (bins) as opposed to total
stats; default=FALSE}
}
\value{
a list containing two tables: 'density' and 'area'
}
\description{
Calculate area and density of cells within a given distance to a tumor boundary
}
