% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{calculateInterfaceStats}
\alias{calculateInterfaceStats}
\title{Calculate area and density of interface}
\usage{
calculateInterfaceStats(aFiles, dataFile, cellTypeFile, cellTypeName,
  fovBoundingBox, pad, plotBoundingBox, funcMarker = NULL,
  writeCSVfiles = TRUE)
}
\arguments{
\item{aFiles}{a vector of Halo boundary annotation files, with 
'.annotations' extension in XML format}

\item{dataFile}{a *.rda file with Halo data for one sample}

\item{cellTypeFile}{file containing cell types to be calculated; each line
is one cell type and each cell type can be either a single
marker name or a comma-separated list of markers}

\item{cellTypeName}{a name to represent the cell types in marker file}

\item{fovBoundingBox}{a list with four elements: X0, X1, Y0, Y1, representing the minimum
and maximum coordinates of the FOV to be plotted}

\item{pad}{amount to trim from FOV before calculating}

\item{plotBoundingBox}{a list with four elements: X0, X1, Y0, Y1, representing the minimum
and maximum coordinates of the outermost plot box [ TO DO: UPDATE THIS ]}

\item{funcMarker}{this marker will be added to all markers in cellTypeFile in order to
compare cells that are +/- for this functional marker [ TO DO: REWORD THIS ]}

\item{writeCSVfiles}{logical indicating whether to write density and area tables to file; 
default=TRUE}
}
\value{
a list containing two tables: 'density' and 'area'
}
\description{
calculate interface stats
}
