% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{calculateMarkerDensityTotalFOV}
\alias{calculateMarkerDensityTotalFOV}
\title{Calculate marker densities for full FOVs}
\usage{
calculateMarkerDensityTotalFOV(dat, areas, markers, writeCSVfiles = TRUE,
  densityDir = NULL)
}
\value{
table of marker densities per FOV
}
\description{
Calculate marker densities for full FOVs
}
