% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cellDive.calculateInfiltrationArea}
\alias{cellDive.calculateInfiltrationArea}
\title{Calculate infiltration area}
\usage{
cellDive.calculateInfiltrationArea(allDat, pp, haloAnnotations,
  updatedExclusions)
}
\arguments{
\item{allDat}{tibble of all halo data}

\item{pp}{list of all project parameters}

\item{haloAnnotations}{list of parsed halo annotations organized by sample and then by FOV}

\item{updatedExclusions}{logical indicating whether exclusions (haloAnnotations) were updated during current run of the pipeline}
}
\value{
list of three items:
          (1) dat = list where area = table of Sample,SPOT,Area and bandAssignments = table of 0/1 for each marker in each cell, plus a column "Band" indicating roughly how far the cell lies from a tumor interface
          (2) updated = logical indicating whether table being returned was generated (TRUE) or loaded from existing file (FALSE)
          (3) pp = updated list of project parameters
}
\description{
Generate table with columns: Sample, SPOT, Band, Area
}
