% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cellDive.calculateInfiltrationDensity}
\alias{cellDive.calculateInfiltrationDensity}
\title{Calculate infiltration density}
\usage{
cellDive.calculateInfiltrationDensity(markerConfig, infiltrationAreas,
  bandAssignments, pp, updatedInfiltrationAreas,
  infiltrationDensityPrefix = "")
}
\arguments{
\item{markerConfig}{parsed marker configuration (as returned by getMarkerConfig())}

\item{infiltrationAreas}{area table containing Sample,SPOT,Band,Area}

\item{bandAssignments}{table containing Band column where each value is an assignment for a single cell to a specific distance band around a tumor interface}

\item{pp}{list of all project parameters}

\item{updatedInfiltrationAreas}{logical indicating whether infiltration areas were updated during the current pipeline run}

\item{infiltrationDensityPrefix}{prefix added to infiltration_density_file/dir in list of project params; leave equal to "" to indicate no prefix added}
}
\value{
list of three items:
          (1) dat = table including Sample,SPOT,Band,Counts,Density 
          (2) updated = logical indicating whether table being returned was generated (TRUE) or loaded from existing file (FALSE)
          (3) pp = updated list of project parameters
}
\description{
Generate table with five columns: Sample,SPOT,Band,Counts,Density
}
