% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cellDive.calculateTotalFOVdensity}
\alias{cellDive.calculateTotalFOVdensity}
\title{Calculate density of markers in each full FOV}
\usage{
cellDive.calculateTotalFOVdensity(allDat, markerConfig, fovAreas, pp,
  updatedFOVarea, updatedComboTable, fovDensityPrefix = "")
}
\arguments{
\item{allDat}{all Halo data}

\item{markerConfig}{tibble of parsed marker configuration (value returned from getMarkerConfig)}

\item{fovAreas}{tibble of areas for each FOV in each Sample (value returned from cellDive.calculateTotalFOVarea)}

\item{pp}{list of all project parameters}

\item{updatedComboTable}{logical indicating whether combo table was updated during current pipeline run}

\item{fovDensityPrefix}{prefix added to prefix added to 'fov_density_file' and/or 'fov_density_dir' in names of pp list (leave "" if no prefix)}

\item{updatedFOVareas}{logical indicating whether FOV areas were updated during current pipeline run}
}
\value{
list of three items:  
          (1) dat = tibble of three columns: Sample, SPOT, Count, Density 
          (2) updated = logical indicating whether a new density table was generated (TRUE) or an old one was loaded (FALSE) 
          (3) pp = updated list of project parameters
}
\description{
Generate table of four columns: Sample, SPOT, Counts, Density
}
