% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cellDive.fillInMarkerComboInterpretations}
\alias{cellDive.fillInMarkerComboInterpretations}
\title{Add columns Cell_Type and Cell_Subtype to marker counts file}
\usage{
cellDive.fillInMarkerComboInterpretations(xlfile = NULL,
  markerComboTable = NULL, outDir = NULL, allCellTypes = NULL)
}
\arguments{
\item{xlfile}{XLSX file containing table described above; default=NULL; if not specified, markerComboTable must be given}

\item{markerComboTable}{marker combination table returned from cellDive.getMarkerCombinationTable(); if not specified, must provide xlfile}

\item{outDir}{directory where xlsx file should be saved}

\item{allCellTypes}{parsed cell types as returned by flattenMetaData()}
}
\value{
marker combo table with columns Cell_Type and Cell_Subtype filled in
}
\description{
Given a table where the first three columns = Markers, Cell_Type, and Cell_Subtype and Cell_Type/Subtype are blank, and the remaining columns contain total counts for each marker combination in a single sample, fill in blank columns using meta data files
}
