% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cellDive.getAllBoundaryAnnotations}
\alias{cellDive.getAllBoundaryAnnotations}
\title{Load and/or parse all Halo boundary annotations}
\usage{
cellDive.getAllBoundaryAnnotations(pp, sampAnn = NULL)
}
\arguments{
\item{pp}{list of all project parameters, from which it will be determined whether annotations have been previously parsed. List must include annotations_dirs and/or annotations_file parameters.}

\item{sampAnn}{parsed sample annotations; only required if pp$annotations_file is NULL or the file does not exist/is empty}
}
\value{
list containing three values: 
          (1) dat = list of all Halo boundary annotations organized by sample and FOV
          (2) updated = logical indicating whether annotations were just parsed (TRUE) or previously parsed (FALSE)
          (3) pp = updated list of project parameters
}
\description{
If boundary annotations have already been parsed and saved to *.rda file, load file. Otherwise, parse all annotations, save to *.rda file and return parsed annotations.
}
