% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cellDive.initCombinedProject}
\alias{cellDive.initCombinedProject}
\title{Process and load any existing project data}
\usage{
cellDive.initCombinedProject(args)
}
\arguments{
\item{pp}{list of project parameters and values}
}
\value{
list of all data loaded, including area, density, project configs and status
}
\description{
Process and load any existing project data in order to start analysis
where previous analyses left off
}
