% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cellDive.loadStudyAnnotations}
\alias{cellDive.loadStudyAnnotations}
\title{Load and parse all Sample+FOV annotations}
\usage{
cellDive.loadStudyAnnotations(metaFiles)
}
\arguments{
\item{metaFiles}{vector of meta data files, including at least
*SampleAnnotations.xlsx and *FOVannotations.xlsx}
}
\value{
tibble of joined Sample+FOV annotations
}
\description{
Load and parse all Sample+FOV annotations
}
