% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cellDive.markExclusions}
\alias{cellDive.markExclusions}
\title{Mark cells in *.rda file to be excluded from analysis}
\usage{
cellDive.markExclusions(projectParams)
}
\arguments{
\item{projectParams}{a list of parameters for entire project including meta directory/files, drift directory, halo annotations directories, and raw data directory}
}
\value{
updated projectParams including data directory where new *.rda files are saved
}
\description{
Add 'EXCLUDE' column to each sample tibble indicating reason(s) why a cell should
be excluded from analysis. Possible reasons include: 
   (1) cell falls outside padding border
   (2) lab determined FOV and/or Marker should be excluded for technical reasons
   (3) drift/loss percentage is above maximum allowed value
   (4) cell was determined by Halo to fall inside an area to be excluded
Write updated tibble to *.rda file with "_Excl.rda" appended to original file name
}
