% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cellDive.updateManifest}
\alias{cellDive.updateManifest}
\title{Overwrite manifest file with updated values}
\usage{
cellDive.updateManifest(projectParams, manifest)
}
\arguments{
\item{projectParams}{key-value pairs of project parameters in list format}

\item{manifest}{file to which parameters should be written in YAML format}
}
\value{
nothing
}
\description{
Overwrite manifest file with most current values
}
