% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cellDive.writeMarkerComboTables}
\alias{cellDive.writeMarkerComboTables}
\title{Write XLSX table of counts for all marker combinations in all samples}
\usage{
cellDive.writeMarkerComboTables(pp, allDat, whichTableFile, markerConfig,
  updatedExclusions)
}
\arguments{
\item{pp}{list of all project parameters}

\item{allDat}{tibble of all Halo data}

\item{whichTableFile}{name of parameter in pp to look at when checking to see if table already exists}

\item{markerConfig}{parsed marker configuration (value returned from getMarkerConfig())}

\item{updatedExclusions}{logical indicating whether exclusions have been updated}
}
\value{
list of three items:  
          (1) dat = list containing one item, the count table generated
          (2) updated = logical indicating whether a new count table was generated (TRUE) or an old one was loaded (FALSE) 
          (3) pp = updated list of project parameters
}
\description{
Write XLSX table of counts for all marker combinations in all samples
}
