% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.R, R/melanoma_spatial.R
\name{cleanBoundaries}
\alias{cleanBoundaries}
\alias{cleanBoundaries}
\title{Remove exclusion boundaries that are contained in another one}
\usage{
cleanBoundaries(boundaries)

cleanBoundaries(boundaries)
}
\arguments{
\item{boundaries}{table generated by readHaloAnnotations}

\item{boundaries}{table generated by readHaloAnnotations}
}
\description{
Remove data from boundaries table that represent exclusion boundaries
that are completely surrounded by another exclusion boundary

Remove data from boundaries table that represent exclusion boundaries
that are completely surrounded by another exclusion boundary
}
