% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{cleanBoundaries_OLD}
\alias{cleanBoundaries_OLD}
\title{Remove exclusion boundaries that are contained in another one}
\usage{
cleanBoundaries_OLD(boundaries)
}
\arguments{
\item{boundaries}{table generated by readHaloAnnotations}
}
\description{
Remove data from boundaries table that represent exclusion boundaries
that are completely surrounded by another exclusion boundary
}
