% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halo.R
\name{cleanData}
\alias{cleanData}
\title{Clean data table}
\usage{
cleanData(dat, v = TRUE, lf = NULL, debug = FALSE)
}
\arguments{
\item{dat}{tibble containing marker data}

\item{v}{verbose - when set to TRUE, messages
will be printed to screen as well as to file;
DEFAULT = TRUE}

\item{lf}{log file}

\item{debug}{print extra messages for debugging}
}
\value{
modified tibble
}
\description{
Alter column names, ensure correct data types for certain columns
}
