% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_combo_table.R
\name{cleanMarkers}
\alias{cleanMarkers}
\title{Clean marker file}
\usage{
cleanMarkers(markerFile, altBases = NULL)
}
\arguments{
\item{markerFile}{File condatining list of marker names}

\item{altBases}{a vector of alternate markers that will be considered the baseline
for all other counts}
}
\value{
a vector of markers
}
\description{
Remove spaces, new lines, tabs from file
}
