% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scheme.R
\name{color_scheme}
\alias{color_scheme}
\title{Color Scheme}
\usage{
color_scheme(scheme_num = 1)
}
\arguments{
\item{scheme_num}{An integer from 1 to 5 indicating the desired color scheme.}
}
\value{
A vector of color codes representing the selected color scheme.
}
\description{
This function returns a predefined color scheme based on the input number.
}
\examples{
\dontrun{
# Load the package
library(your_package_name)

# Get color scheme 3
colors <- color_scheme(3)
}
}
