% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{computeBandAreasMC}
\alias{computeBandAreasMC}
\title{Compute band areas for each interface bin}
\usage{
computeBandAreasMC(tumorBoundariesList, bb, interfaceBins, nGrid = 100,
  maxSeg = 1000, exclusionB)
}
\arguments{
\item{tumorBoundariesList}{a list where each element is a tibble containing tumor boundaries
returned from readHaloAnnotations()}

\item{bb}{a list containing X0,X1,Y0,Y1 representing the boundary
of the trimmed FOV}

\item{interfaceBins}{vector of distances (integers) that define each band; default = (-20:20)*10}

\item{nGrid}{number of random points to use for area calculations; default=100}

\item{maxSeg}{maximum segments to divide long boundary segments into; default=1000}

\item{exclusionsB}{a list where each element is a tibble containing exclusion boundaries
returned from readHaloAnnotations()}
}
\description{
Given a list of distances from a tumor boundary, or 'interface bins', calculate area
for each bin using a sobol grid
}
