% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_meta.R
\name{configureStudy}
\alias{configureStudy}
\title{Configure study}
\usage{
configureStudy(studyName = NULL, studyDir = NULL,
  setDefaultDirectoryStruct = TRUE, configDir = NULL,
  studyConfigFile = NULL, raw_data_dir = NULL, raw_data_files = NULL,
  data_dir = NULL, data_files = NULL, meta_dir = NULL,
  meta_files = NULL, annotations_dirs = NULL, annotations_files = NULL,
  drift_dir = NULL, drift_files = NULL, study_dir = NULL,
  cohort_dir = NULL, infiltration_dir = NULL, fov_stats_dir = NULL,
  infiltration_density_dir = NULL, infiltration_density_file = NULL,
  infiltration_area_dir = NULL, infiltration_band_assignments_file = NULL,
  infiltration_area_file = NULL, fov_density_dir = NULL,
  fov_density_file = NULL, fov_area_dir = NULL, fov_area_file = NULL,
  marker_config_file = NULL, plot_config_file = NULL,
  marker_analysis_config_file = NULL, celltype_config_file = NULL,
  log = NULL, debug = TRUE, raw_marker_combo_table_file = NULL,
  pad = 20, drift_threshold = 0.1, updateExistingConfigFiles = TRUE,
  write_csv_files = TRUE, max_g = 5, band_width = 10,
  maximumDistanceFromInterface = 360)
}
\value{
nothing
}
\description{
Set study parameters and write YAML file to be used as study config
}
