% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.R
\name{convertBoundaryFilePaths}
\alias{convertBoundaryFilePaths}
\title{Convert shared drive path to server path, and add subdirectory for exclusion coordinates}
\usage{
convertBoundaryFilePaths(paths, subdir, from = ".*bic.mskcc.org",
  to = "/ifs/tcga/socci/Multiomyx")
}
\arguments{
\item{paths}{a vector of paths to be converted}

\item{from}{string to be replaced}

\item{to}{string to replace {from}}
}
\value{
a vector equal in length to input vector, with paths converted and ExclusionCoordinates
        subfolder added
}
\description{
Do a simple string substitution to convert one path to another, and add exclusion 
coordinates subfolder to all paths
}
