% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_util.R
\name{convertCellMinMaxToMidpoints}
\alias{convertCellMinMaxToMidpoints}
\title{Replace Min/Max X and Y coordinates from Halo with cell midpoints}
\usage{
convertCellMinMaxToMidpoints(ds)
}
\arguments{
\item{ds}{tibble containing columns XMax, Xmin, YMax, YMin, Sample,
SPOT, UUID, Marker, Value}
}
\value{
tibble containing only those columns mentioned above
}
\description{
Given a tibble containing columns XMax, Xmin, YMax, YMin, replace
those columns with 'X' and 'Y' columns which contain the midpoint
between those min and max values
}
