% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{densityBoxPlots}
\alias{densityBoxPlots}
\title{Box plots of densities showing differences between samples for different cell types}
\usage{
densityBoxPlots(markers, density, sampleOrder = NULL, starSize = 8,
  pdfFile = NULL)
}
\arguments{
\item{markers}{vector of markers, for each of which a plot will be generated}

\item{density}{tibble of density data, including columns for at least Sample, CellType
and [Total|Density]}

\item{sampleOrder}{if specified, sample boxes will appear in this order; default=NULL}

\item{starSize}{text size, specifically for significance starts; default=8}

\item{pdfFile}{if specified, plots will be printed to this file; default=NULL}
}
\value{
nothing
}
\description{
Box plots of densities showing differences between samples for different cell types
}
