% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_meta.R
\name{flattenMetaData}
\alias{flattenMetaData}
\title{Flatten meta data for a single study}
\usage{
flattenMetaData(metaDir = NULL, metaFiles = NULL, sampAnnFile = NULL,
  fovAnnFile = NULL, cellTypesFile = NULL, haloPanelsFile = NULL,
  markerDescFile = NULL)
}
\arguments{
\item{metaDir}{directory containing all available meta data files in XLSX format; default=NULL}

\item{metaFiles}{vector of meta data files; an alternative to providing metaDir; default=NULL}

\item{sampAnnFile}{sample annotation file in XLSX format; if given, this value will override a sample
annotation file found in metaDir or metaFiles; default=NULL}

\item{fovAnnFile}{FOV annotation file in XLSX format; if given, this value will override a FOV
annotation file found in metaDir or metaFiles; default=NULL}

\item{cellTypesFile}{XLSX file describing all known cell types (marker combinations); if given, this 
value will override a cell types file found in metaDir or metaFiles; default=NULL}

\item{haloPanelsFile}{XLSX file describing all halo panels; if given, this value will override a halo
panels file found in metaDir or metaFiles; default=NULL}

\item{markerDescFile}{XLSX file describing individual markers; if given, this value will override a 
marker description file found in metaDir or metaFiles; default=NULL}

\item{unreasonalbesFile}{#### MOST LIKELY WILL BE REMOVED}
}
\value{
a list containing three data frames: SlideAnnotation, MarkerAnnotation and MarkerCombinationAnnotation
}
\description{
Given all available meta data files for a single study, flatten all data into three
tables: 
  (1) SlideAnnotation - all information regarding samples and FOV
  (2) MarkerAnnotation - all information describing individual markers
  (3) MarkerCombinationAnnotation - all information describing important marker
      combinations
}
