% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{g_legend}
\alias{g_legend}
\title{Get legend from a gplot}
\usage{
g_legend(plt)
}
\arguments{
\item{a.gplot}{a plot generated by ggplot()}
}
\value{
the legend of a.gplot
}
\description{
Given a plot generated with ggplot(), separate and return only
the legend component
}
