% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{generateSobolGrid}
\alias{generateSobolGrid}
\title{Generate data frame of random points (is it random???)}
\usage{
generateSobolGrid(nGrid, bbG)
}
\arguments{
\item{nGrid}{number of points to generate}

\item{bbG}{a list containing X0,X1,Y0,Y1 representing the boundary
of the plot area}
}
\value{
data frame of X and Y values of the random points
}
\description{
Generate a data frame of random points that fall within the plot area
}
