% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_custom_colors}
\alias{get_custom_colors}
\title{Get marker colors}
\usage{
get_custom_colors(markerColors = NULL, markerCombos = NULL)
}
\arguments{
\item{markerColors}{named vector of colors already in use, where names are 
marker names (e.g., CD3, CD4, MRC1, etc.) and values are
colors; only to be used when 'custom' is set to FALSE; default=NULL}

\item{markerCombos}{vector of marker combinations to add to MarkerColors; only to be
used when 'custom' is set to FALSE}
}
\description{
Get a named vector of predefined colors for certain markers
}
