% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{get_HaloDotAPI_token}
\alias{get_HaloDotAPI_token}
\alias{set_HaloDotAPI_token}
\title{Halo Infinite API Token Management}
\usage{
get_HaloDotAPI_token()

set_HaloDotAPI_token(token = NULL)
}
\arguments{
\item{token}{HaloDotAPI Token.}
}
\value{
\code{get_HaloDotAPI_token()} invisibly returns the API token if it is
found as an environment variable.
}
\description{
Get or set your Halo Infinite API access token as an environment variable.
Calling \code{set_HaloDotAPI_token()} without the \code{token} argument will allow
the user to provide their token in a pop-up window using
\code{askpass::askpass()} to avoid entering the token into the \code{R} terminal.
}
