% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\docType{data}
\name{get_HaloDotAPI_version}
\alias{get_HaloDotAPI_version}
\alias{set_HaloDotAPI_version}
\alias{HaloDotAPI_versions}
\title{Halo Infinite API Version}
\format{
An object of class \code{character} of length 16.
}
\usage{
get_HaloDotAPI_version()

set_HaloDotAPI_version(version)

HaloDotAPI_versions
}
\arguments{
\item{version}{HaloDotAPI Version.}
}
\value{
\code{get_HaloDotAPI_version()} returns the API version if it is
found as an environment variable.
}
\description{
Get or set your Halo Infinite API version as an environment variable.
}
\details{
Unless otherwise specified, all of the requesting functions will,
by default, use the most recent version of the Halo Infinite API as listed
in \code{HaloDotAPI_versions}.
}
\keyword{datasets}
