% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_marker_colors}
\alias{get_marker_colors}
\title{Get marker colors}
\usage{
get_marker_colors(markerColors = NULL, markerCombos = NULL,
  custom = FALSE)
}
\arguments{
\item{markerColors}{named vector of colors already in use, where names are 
marker names (e.g., CD3, CD4, MRC1, etc.) and values are
colors; only to be used when 'custom' is set to FALSE; default=NULL}

\item{markerCombos}{vector of marker combinations to add to MarkerColors; only to be
used when 'custom' is set to FALSE}

\item{custom}{logical, default=FALSE; use custom colors previously 
defined as: 
  CD3,CD4,CD8,FOXP3,CD56: blues
  CD68, CD14,CD163,MRC1,TGM2: reds
  SOX10: yellow
  CD20: green
  other combinations: white}
}
\description{
Get a named vector of predefined colors for certain markers
}
