% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_meta.R
\name{getAllCellTypeCombinations}
\alias{getAllCellTypeCombinations}
\title{Get all possible marker combinations for a certain cell type/subtype}
\usage{
getAllCellTypeCombinations(cellTypeTable, cellType, cellSubtype = NULL)
}
\arguments{
\item{cellTypeTable}{parsed table of all marker combinations, cell types and subtypes}

\item{cellType}{the cell type to filter for}

\item{cellSubtype}{the cell subtype to filter for}
}
\value{
a vector of cell type marker combinations
}
\description{
Create list of all marker combinations that define a given cell type and/or
subtype
}
