% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getAllCombos}
\alias{getAllCombos}
\title{Get all possible combinations of markers}
\usage{
getAllCombos(markers)
}
\arguments{
\item{markers}{vector of single markers}
}
\value{
list of all combinations
}
\description{
Given a vector of single markers, return a list
of all possible combinations of those markers,
including both positive and negative variations
}
