% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{getAllFOVDensityValues}
\alias{getAllFOVDensityValues}
\title{Get all absolute and percentage density values for complete FOVs}
\usage{
getAllFOVDensityValues(den, areaDat)
}
\value{
table of all density values
}
\description{
Given FOV-level density values, add sample-level values to data frame by  
calculating total densities of each marker in ALL FOV for a single sample
}
