% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{getAllInfiltrationPlots}
\alias{getAllInfiltrationPlots}
\title{Create and store all infiltration plots in a list to be printed later}
\usage{
getAllInfiltrationPlots(den, bandWidth = NULL, config,
  yScaleConsistency = "population", absoluteDensity = TRUE,
  densityPercentage = TRUE, byFOV = TRUE, summarize = TRUE,
  stacked = TRUE, sampleOrder = NULL, infiltrationAreas = NULL,
  indivPopulations = TRUE, forceStack = FALSE)
}
\arguments{
\item{den}{tibble with columns for Sample,SPOT,Band,Counts,Density}

\item{bandWidth}{width of each band around tumor interface}

\item{config}{parsed marker config (return value from getMarkerConfig())}

\item{yScaleConsistency}{possible values: "population", "all", or "sample" indicating which
subset of values to use to set maximum for all plots}

\item{absoluteDensity}{logical indicating whether to plot density values on absolute scale (default=TRUE)}

\item{densityPercentage}{logical indicating whether to plot density percentages (default=TRUE)}

\item{byFOV}{logical indicating whether to generate plots for each individual FOV (default=TRUE)}

\item{summarize}{logical indicating whether to generate summary plots of total band densities (ALL FOVs; default=TRUE)}

\item{stacked}{logical indicating whether to generate plots of all cell types for a single marker set
in a stacked bar chart (default=TRUE)}

\item{sampleOrder}{vector of sample names in order they should appear on plots}

\item{infiltrationAreas}{area data for each sample/spot/band}

\item{indivPopulations}{plot populations individually; default=TRUE; if set to FALSE, ONLY stacked plots
will be created}
}
\value{
list of all plots
}
\description{
Create and store all infiltration plots in a list to be printed later
}
