% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{getAllTotalFOVPlots}
\alias{getAllTotalFOVPlots}
\title{Create and store all infiltration plots in a list to be printed later}
\usage{
getAllTotalFOVPlots(den, config, indivPopulations = TRUE,
  absoluteDensity = TRUE, densityPercentage = TRUE, summarize = TRUE,
  stacked = TRUE, fovAreas = NULL, forceStack = FALSE, yMax = NULL,
  annotationCols = NULL)
}
\arguments{
\item{den}{tibble with columns for Sample,SPOT,Band,Counts,Density}

\item{config}{parsed marker config (return value from getMarkerConfig())}

\item{indivPopulations}{plot populations individually; default=TRUE; if set to FALSE, ONLY stacked plots
will be created}

\item{absoluteDensity}{logical indicating whether to plot density values on absolute scale (default=TRUE)}

\item{densityPercentage}{logical indicating whether to plot density percentages (default=TRUE)}

\item{summarize}{logical indicating whether to generate summary plots of total band densities (ALL FOVs; default=TRUE)}

\item{stacked}{logical indicating whether to generate plots of all cell types for a single marker set
in a stacked bar chart (default=TRUE)}

\item{fovAreas}{area data for each sample/spot}

\item{forceStack}{force stacking of cell types even if they appear to be unstackable}

\item{yMax}{maximum y value for all plots}

\item{annotationCols}{vector of column names to use for annotation}

\item{sampleOrder}{vector of sample names in order they should appear on plots}
}
\value{
list of all plots
}
\description{
Create and store all infiltration plots in a list to be printed later
}
