% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{getAreaPerBand}
\alias{getAreaPerBand}
\title{Compute lattice of boundary bins}
\usage{
getAreaPerBand(allBoundaries, sample, fov, maxG, haloInfiltrationDir,
  outDir = NULL, bbData = NULL, writeCSVfiles = FALSE,
  interfaceBins = NULL)
}
\arguments{
\item{sample}{sample name}

\item{fov}{FOV}

\item{maxG}{[TO DO: FIND OUT WHAT THIS IS]}

\item{haloInfiltrationDir}{directory containing halo infiltration files with suffix 
"_infiltration_ex_data"}

\item{outDir}{if writeCSVfiles=TRUE, files will be written here}

\item{bbData}{a list containing X0,X1,Y0,Y1 representing the boundary
of the trimmed FOV}

\item{writeCSVfiles}{write lattice band area file in csv format; default=FALSE}

\item{interfaceBins}{vector of distances (integers) that define each band; default = (-20:20)*10}

\item{aFile}{Halo boundary annotation file in XML format}
}
\value{
tibble
}
\description{
Given an annotation file and halo infiltration data, compute tibble of Sample, FOV, 
Halo Version, MaxG, Band, Area, Halo
}
\details{
*NOTE: formerly computeBoundaryBinsLattice
}
