% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaceTools.R
\name{getBandCounts}
\alias{getBandCounts}
\title{Get counts of cells that fall within given distances of
tumor boundaries for a single FOV}
\usage{
getBandCounts(ds, bbData, bbPlot, bbFOV0, aFile, doPlot = T,
  interfaceBands = NULL)
}
\arguments{
\item{ds}{a tibble containing data for one FOV, including columns
for X, Y, and one for each marker}

\item{bbData}{a list containing X0,X1,Y0,Y1 representing the boundary
of the trimmed FOV}

\item{bbPlot}{a list containing X0,X1,Y0,Y1 to be used as boundary of
entire plot}

\item{bbFOV0}{a list containing X0,X1,Y0,Y1 showing the actual boundary
of the FOV (untrimmed)}

\item{aFile}{Halo boundaries annotations file in XML format}

\item{doPlot}{logical indicating whether to plot cell type locations; 
default=TRUE}

\item{interfaceBands}{vector of distances}
}
\description{
Given a set of interface bands, or set distances from a tumor 
boundary, count the numbers of cells that fall within those bands
}
