% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.R
\name{getBorderPaddingExclusions}
\alias{getBorderPaddingExclusions}
\title{Get indices of cell to be excluded because they lie outside the border padding}
\usage{
getBorderPaddingExclusions(fov, samp, dat, borderPad)
}
\arguments{
\item{fov}{fov to be marked for exclusions}

\item{samp}{sample to be marked}

\item{dat}{data tibble containing data to be marked}

\item{borderPad}{number in microns to trim from data}
}
\description{
Get indices of cell to be excluded because they lie outside the border padding
}
