% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{getBoundingBoxL}
\alias{getBoundingBoxL}
\title{Get min and max X and Y coordinates of a FOV}
\usage{
getBoundingBoxL(spObj)
}
\arguments{
\item{spObj}{spot object(?); data frame containing X and Y coordinates
of all cells in a FOV}
}
\value{
list where X0 = minimum X, X1 = maximum X, and same for Y
}
\description{
Get min and max X and Y coordinates of a FOV
}
