% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_meta.R
\name{getCellTypes}
\alias{getCellTypes}
\title{Make list of all known cell types (marker combinations) given in meta data files}
\usage{
getCellTypes(simpleTypes, complexTypes)
}
\arguments{
\item{simpleTypes}{straight-forward markercombination-cellType pairs}

\item{complexTypes}{table describing combinations that are not exactly straight-forward
(described in meta data format documentation)}

\item{unreasonableCombinations}{**NOTE: WILL MOST LIKELY BE ELIMINULLTED**}
}
\value{
list of two elements: CellTypes and Conditional, where CellTypes are ALL known 
          markercombination-cellType pairs and Conditional is a vector describing cell types
          that are conditional based on the data itself
}
\description{
Make list of all known cell types (marker combinations) given in meta data files
}
