% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.R
\name{getHaloExclusions}
\alias{getHaloExclusions}
\title{Remove points from data set that fall inside exclusion boundaries}
\usage{
getHaloExclusions(fov, dd, excType, haloAnnotations = NULL, aFile = NULL,
  exc = NULL, boundaryColors = NULL, boundaryReassignmentFile = NULL)
}
\arguments{
\item{aFile}{Halo boundaries annotation file in XML format; if NULL, must provide
the list of excluded region tables as returned by cleanBoundaries()}

\item{ds}{tibble containing at minimum, Sample, SPOT, UUID, X, Y, Marker, Value}

\item{excB}{a list of excluded region tables as returned by cleanBoundaries;
if NULL, must provide the original Halo annotations file in XML format}
}
\value{
tibble filtered to remove points that fall inside exclusion boundaries
}
\description{
Read Halo boundaries annotations file and remove from data set any cells
that fall inside the exclusion regions
}
