% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{getInfiltrationDensitySummary}
\alias{getInfiltrationDensitySummary}
\title{Summarize infiltration by getting total density of all FOV
in each band}
\usage{
getInfiltrationDensitySummary(markers, ia = NULL, ba = NULL,
  areaFiles = NULL)
}
\arguments{
\item{markers}{vector of markers for which density is to be calculated}

\item{ia}{tibble of interface area; required if areaFiles not given; default=NULL}

\item{ba}{tibble containing band assignments of markers; required if areaFiles 
not given; default=NULL}

\item{areaFiles}{vector of files containing area and band assignments for each sample; 
required if ia or ba is not given; default=NULL}
}
\value{
tibble containing total density for each band
}
\description{
Summarize infiltration by getting total density of all FOV
in each band
}
