% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_combo_table.R
\name{getMarkerCombos}
\alias{getMarkerCombos}
\title{Build a list of all marker combinations described in "complex" combination
format in cell types spreadsheet (described in docs)}
\usage{
getMarkerCombos(comboType, markers)
}
\arguments{
\item{comboType}{tells what kind of combinations to build ("ALL" indicates
all markers given must stay together, "ANY" means any combination
of given markers, including all different lengths, and 
types like ">2" or "<=3" tells how many markers each combination must 
contain}

\item{markers}{vector of markers to combine}
}
\value{
a list of all possible combinations matching criteria in comboType
}
\description{
Based on combination type ("ALL", "ANY", "<2", ">3", etc.), generate all
possible combinations of markers given
}
