% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getMarkerConfig}
\alias{getMarkerConfig}
\title{Parse marker config}
\usage{
getMarkerConfig(configFile, plotConfigFile)
}
\arguments{
\item{configFile}{configuration file in yaml format}
}
\value{
a tibble containing all configuration information
}
\description{
Given a marker configuration file in yaml format, parse 
data into a tibble
}
