% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{getMarkerDensityTable}
\alias{getMarkerDensityTable}
\title{Get a table of marker combination densities}
\usage{
getMarkerDensityTable(markers, dat = NULL, outDir = getwd(),
  dataFiles = NULL, annotationsDirs = NULL, haloAnnotations = NULL,
  areaFiles = NULL, densityFiles = NULL, sampleOrder = NULL,
  writeCSVfiles = FALSE, pad = 20, byBand = TRUE,
  maxDistanceFromInterface = 360, bandWidth = 10, markerSetName = NULL,
  sortByMarker = NULL, funcMarker = NULL, maxG = 5, calcFOVarea = FALSE)
}
\arguments{
\item{dat}{tibble containing halo data, specifically Sample, UUID, SPOT, Marker,
Xmin/max, Ymin/max, ValueType, Value}

\item{outDir}{if writeCSVfiles=T, directory to which these files will be written}

\item{dataFiles}{vector of *.rda files containing Halo data for a single sample}

\item{annotationsDirs}{vector of directories containing *.annotations files, XML files of boundary
annotations from Halo; order must correspond to order of dataFiles}

\item{writeCSVfiles}{logical; write CSV files, one of density values and one of area values; 
default=TRUE}

\item{pad}{amount to trim from each FOV}

\item{byBand}{logical indicating whether to break down density by distance from tumor interface; 
default=FALSE}

\item{maxDistanceFromInterface}{include only those cells within this distance from tumor interface}

\item{bandWidth}{width of each band around tumor interface, to be used when byBand is TRUE; 
default=10}

\item{sortByMarker}{sort FOVs by density of this marker (cell type)in ascending order;
default=NULL}

\item{funcMarker}{character string of functional marker to plot on top of every other marker}

\item{maxG}{maximum factor of 10 to use for generating random points on grid [ REWORD??? ]}
}
\value{
tibble containing density for all markers in all samples
}
\description{
Generate a table of cell type densities
}
