% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{getPlotTheme}
\alias{getPlotTheme}
\title{Get plot theme for a given type of halo plot}
\usage{
getPlotTheme(plotType)
}
\arguments{
\item{plotType}{character string indicating which type of plot youre making;
options: ["infiltration density"]}
}
\value{
theme() object with standard theme elements
}
\description{
Depending on what type of plot you're making, get a theme()
object with the elements standard to that plot type
}
