% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{getSampleBandMaxY}
\alias{getSampleBandMaxY}
\title{Get maximum density value from all Sample+Band combos in a data set}
\usage{
getSampleBandMaxY(den, markers, ia)
}
\arguments{
\item{den}{density tibble}

\item{markers}{markers to consider}

\item{ia}{infiltration area tibble (Sample,Band,Area)}
}
\value{
double; maximum density value from all Sample+band groups
}
\description{
Get maximum density value from all Sample+Band combos in a data set
}
