% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getSampleFromFileName}
\alias{getSampleFromFileName}
\title{Extract sample name from file name}
\usage{
getSampleFromFileName(filePath)
}
\arguments{
\item{filePath}{file, which may or may not include full path}
}
\description{
Assuming sample name does not include underscores,
return everything preceding the first underscore
}
