% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_counts.R
\name{getSheetOrder}
\alias{getSheetOrder}
\title{Get order of workbook sheets for final excel file}
\usage{
getSheetOrder(runCounts = TRUE, runMedians = TRUE, runFracTotal = FALSE,
  altBases = NULL, debug = FALSE)
}
\arguments{
\item{runCounts}{include raw counts in final output; DEFAULT=TRUE}

\item{runMedians}{include a sheet of medians for each counts and fraction sheet; 
DEFAULT=TRUE}

\item{runFracTotal}{include a sheet of fractions of total cells; DEFAULT=FALSE}

\item{altBases}{include a sheet of fractions of cells positive for each of these 
markers; DEFAULT=NULL}

\item{debug}{print debug messages; DEFAULT=FALSE}
}
\description{
Order sheets according to whether fractions and/or medians are to be run
}
