% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.R
\name{getStudyExclusions}
\alias{getStudyExclusions}
\title{Get indices of cells to be excluded because of reasons determined manually by a lab member}
\usage{
getStudyExclusions(fovAnnotations, dat)
}
\arguments{
\item{fovAnnotations}{tibble of FOV annotations for one sample, read from xlsx file
included in study meta data}

\item{dat}{full data set to be marked for exclusions}
}
\value{
vector of dat indices to be marked for exclusion
}
\description{
Get indices of cells to be excluded because of reasons determined manually by a lab member
}
