% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_meta.R
\name{getTemplatePlotConfig}
\alias{getTemplatePlotConfig}
\title{Generate a template plot configuration based on marker configuration}
\usage{
getTemplatePlotConfig(markerConfig, writeYAML = TRUE, outDir = getwd())
}
\arguments{
\item{markerConfig}{marker configuration in nested list format (most likely previously-read
from a YAML file); if NULL, user must provide a marker config file in YAML
format; default=NULL}

\item{writeYAML}{logical indicating whether to write configuration to file; default=TRUE}

\item{outDir}{directory where YAML file should be written; default=getwd()}

\item{markerConfigFile}{marker configuration YAML file; if NULL, user must provide configuration
in list format; default=NULL}
}
\value{
all plot information and default plot configuration values
}
\description{
Given either a marker configuration in list format, or a configuration file in YAML
format, create a template plot configuration to be modified manually by user for plotting
}
