% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_meta.R
\name{getTemplateStudyConfig}
\alias{getTemplateStudyConfig}
\title{Create a study config template}
\usage{
getTemplateStudyConfig(studyDirectory = getwd(),
  setDefaultDirectoryStruct = TRUE, studyName = NULL, writeYAML = TRUE)
}
\arguments{
\item{setDefaultDirectoryStruct}{}
}
\description{
Create a study config template, returning config in list format and optionally writing
config to yaml file. User may choose to set all parameters to NULL or to set any available
default values.
}
