% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{getTotalDensityFillColors}
\alias{getTotalDensityFillColors}
\title{Get a named vector of colors to use on total density bar chart}
\usage{
getTotalDensityFillColors(dat, funcPosColor, funcNegColor, sampleColors,
  sampleColorsFuncPos)
}
\arguments{
\item{dat}{tibble including at least columns for Sample,SPOT,Functional}

\item{funcPosColor}{color to represent density of cells positive for a functional marker}

\item{funcNegColor}{color to represent density of cells negative for a functional marker}

\item{sampleColors}{vector of colors, each named according to the sample it represents}

\item{sampleColorsFuncPos}{vector of colors for density of cells positive for functional marker, 
named according to the samples they represent}
}
\value{
vector of colors, also named by those colors
}
\description{
Using a tibble including at least columns for Sample,SPOT,Functional, 
create a named vector to be added to that tibble and used to fill
bars on a chart for total density
}
