% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{getTotalDensityPlot}
\alias{getTotalDensityPlot}
\title{Build bar plot for total cell type density}
\usage{
getTotalDensityPlot(dat, onePanel = TRUE, funcMarker = NULL,
  funcPosColor = "grey35", funcNegColor = "grey50",
  sampleColor = "#f16913", sampleColorDark = "#80380A",
  consistentYScaling = TRUE)
}
\arguments{
\item{dat}{tibble containing all info needed for plotting [ TO DO: FILL IN EXACTLY WHAT'S NEEDED ]}

\item{onePanel}{logical; place all plots on one panel}

\item{funcMarker}{marker to be indicated as either positive or negative for each cell type}

\item{funcPosColor}{hex color indicating functional marker is positive in that cell type}

\item{funcNegColor}{hex color indicating functional marker is negative in that cell type}

\item{sampleColor}{vector of colors named by the sample which it will represent; this color will only
appear on the median bars}

\item{sampleColorDark}{vector of colors named by sample, to represent density of cells that are positive
for the functional marker, if one is given}

\item{consistentYScaling}{logical; when onePanel=FALSE, set this parameter to FALSE in order to scale separate 
plots individually; default=TRUE}
}
\value{
list of ggplots to be printed
}
\description{
Build bar plot for total cell type density
}
