% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appearance.R
\name{halo_appearance}
\alias{halo_appearance}
\title{Players Appearance on Halo Infinite}
\usage{
halo_appearance(gamertag, version = "0.3.6", token = get_HaloDotAPI_token())
}
\arguments{
\item{gamertag}{Player's Gamertag.}

\item{version}{HaloDotAPI Version.}

\item{token}{HaloDotAPI Token.}
}
\value{
A \code{list} with class \code{appearance}.
}
\description{
Returns targeted player's appearance on Halo Infinite.
}
