% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\name{halo_articles}
\alias{halo_articles}
\title{Halo Infinite Articles}
\usage{
halo_articles(
  language = "en-US",
  version = get_HaloDotAPI_version(),
  token = get_HaloDotAPI_token()
)
}
\arguments{
\item{language}{Desired language. One of \code{'de-DE'}, \code{'es-ES'},
\code{'es-MX'}, \code{'en-US'}, \code{'fr-FR'}, \code{'it-IT'}, \code{'ja-JP'}, \code{'ko-KR'},
\code{'pl-PL'}, \code{'pt-BR'}, \code{'ru-RU'}, \code{'zh-CN'}, or \code{'zh-HK'}.
Default is \code{'en-US'}.}

\item{version}{HaloDotAPI Version.}

\item{token}{HaloDotAPI Token.}
}
\value{
A \code{list} with class \code{articles}.
}
\description{
Returns available articles on Halo Infinite.
}
