% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{halo_matches_retrieve}
\alias{halo_matches_retrieve}
\title{Halo Infinite Match Details}
\usage{
halo_matches_retrieve(
  id,
  version = get_HaloDotAPI_version(),
  token = get_HaloDotAPI_token(),
  .progress = TRUE
)
}
\arguments{
\item{id}{Character vector of match IDs.}

\item{version}{HaloDotAPI Version.}

\item{token}{HaloDotAPI Token.}

\item{.progress}{Logical. Should a progress bar be displayed? Passed on to
\code{furrr::future_map()}.}
}
\value{
A \code{list} with class \code{stats_matches_retrieve}.
}
\description{
Returns targeted match's details on Halo Infinite.
}
