% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medals.R
\name{halo_medals}
\alias{halo_medals}
\title{Halo Infinite Medals}
\usage{
halo_medals(version = "0.3.6", token = get_HaloDotAPI_token())
}
\arguments{
\item{version}{HaloDotAPI Version.}

\item{token}{HaloDotAPI Token.}
}
\value{
A \code{list} with class \code{metadata_medals_list}.
}
\description{
Returns available medals on Halo Infinite.
}
