% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugc.R
\name{halo_UGC}
\alias{halo_UGC}
\title{Halo Infinite User-Generated Content}
\usage{
halo_UGC(
  count = NULL,
  offset = NULL,
  author = NULL,
  kind = NULL,
  tags = NULL,
  version = "0.3.6",
  token = get_HaloDotAPI_token()
)
}
\arguments{
\item{count}{Number of items to retrieve.}

\item{offset}{Number of items to offset.}

\item{author}{Asset's author (i.e. gamertag).}

\item{kind}{Asset's kind.}

\item{tags}{Character vector of tags.}

\item{version}{HaloDotAPI Version.}

\item{token}{HaloDotAPI Token.}
}
\value{
A \code{list} with class \code{ugc_search}.
}
\description{
Search for UGC (User-Generated Content) items on Halo Infinite.
}
