% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaloDotAPI.R
\name{HaloDotAPI}
\alias{HaloDotAPI}
\title{Perform a Request to the Halo Infinite API}
\usage{
HaloDotAPI(
  endpoint,
  ...,
  version = get_HaloDotAPI_version(),
  token = get_HaloDotAPI_token()
)
}
\arguments{
\item{endpoint}{Endpoint.}

\item{...}{Query parameters passed on to the chosen \code{endpoint}.}

\item{version}{HaloDotAPI Version.}

\item{token}{HaloDotAPI Token.}
}
\value{
A \code{list} with a class that is specific to the chosen \code{endpoint}.
}
\description{
This is a low-level function for making requests to the Halo Infinite
API. Most users will not need to use this function directly, but will
instead use one of the wrapper functions that are specific to the
available endpoints.
}
