% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{HaloDotAPI_versions}
\alias{HaloDotAPI_versions}
\title{Halo Infinite API Versions}
\format{
A character vector.
}
\usage{
HaloDotAPI_versions
}
\description{
Character vector of available versions for the Halo Infinite API.
}
\examples{
HaloDotAPI_versions

}
\keyword{datasets}
