% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{joinBandArea}
\alias{joinBandArea}
\title{Add columns for band area to existing tibble with a column of bands}
\usage{
joinBandArea(bdat, bandArea)
}
\arguments{
\item{bdat}{tibble of band data, including at least a column named Band}

\item{bandArea}{a data frame or tibble with an area value for each band in bdat}
}
\value{
tibble containing all data in bdat and bandArea
}
\description{
Add columns for band area to existing tibble with a column of bands
}
