% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_util.R
\name{joinBoundingBoxes}
\alias{joinBoundingBoxes}
\title{Get X and Y coordinates of box enclosing two existing boxes}
\usage{
joinBoundingBoxes(b1, b2)
}
\description{
Get X and Y coordinates of box enclosing two existing boxes
}
