% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halo.R
\name{loadHalo}
\alias{loadHalo}
\title{Load data from *.rda file}
\usage{
loadHalo(cancerType, sample, FOV = NULL)
}
\arguments{
\item{cancerType}{cancer type (default: melanoma)}

\item{sample}{sample name, as it is called in *.rda file}

\item{FOV}{(SPOT)    integer; FOV/SPOT number (default: NULL)}
}
\value{
tibble saved in *.rda file
}
\description{
Load data for one subsample from one cancer type, and optionally
one FOV. Default is to load ALL FOV for a subsample
}
