% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{logMsg}
\alias{logMsg}
\title{Log message to file and screen if specified}
\usage{
logMsg(msg, v = TRUE, logFile = NULL, type = "INFO")
}
\arguments{
\item{msg}{message to be printed}

\item{v}{verbose - when set to TRUE (default), messages
will be printed to screen as well as to file}

\item{logFile}{file to write log messages to; DEFAULT=NULL}

\item{type}{message type (INFO, DEBUG, WARN); DEFAULT=INFO}
}
\description{
Print date() and log message to file, and if
verbose is set to TRUE, also print to screen
}
