% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{logParams}
\alias{logParams}
\title{Log project parameters}
\usage{
logParams(project_params, action)
}
\arguments{
\item{project_params}{list of project parameters and values 
generated by projectParams()}

\item{action}{action being carried out for the project (e.g., 
"generating counts" or "making pie charts")}
}
\description{
Log all parameters set using either command line arguments
or manifest file
}
