% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{makeSampleInfiltrationDensityPlots}
\alias{makeSampleInfiltrationDensityPlots}
\title{Generate all infiltration density plots for a single sample}
\usage{
makeSampleInfiltrationDensityPlots(den, config, markers, bandWidth,
  sampleOrder = NULL, plotTitle = NULL, yMax = NULL, sumYmax = NULL,
  byFOV = TRUE, absoluteDensity = TRUE, densityPercentage = TRUE,
  summarize = TRUE, summaryTitle = NULL, forceStack = FALSE)
}
\arguments{
\item{den}{tibble with columns for Sample,SPOT,Band,Counts,Density}

\item{config}{parsed marker config (return value from getMarkerConfig())}

\item{markers}{vector of marker combinations to plot}

\item{bandWidth}{width of each band around tumor interface}

\item{sampleOrder}{vector of sample names in order they should appear on plots}

\item{plotTitle}{title to appear on plot}

\item{yMax}{maximum y value for all plots}

\item{byFOV}{logical indicating whether to generate plots for each individual FOV (default=TRUE)}

\item{absoluteDensity}{logical indicating whether to plot density values on absolute scale (default=TRUE)}

\item{densityPercentage}{logical indicating whether to plot density percentages (default=TRUE)}

\item{summarize}{logical indicating whether to generate summary plots of total band densities (ALL FOVs; default=TRUE)}

\item{summaryTitle}{title to appear on summary plots}

\item{stacked}{logical indicating whether to generate plots of all cell types for a single marker set
in a stacked bar chart (default=TRUE)}
}
\value{
list of sample plots
}
\description{
Generate all infiltration density plots for a single sample
}
