% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_combo_table.R
\name{markerComboCounts}
\alias{markerComboCounts}
\title{Find all marker combinations existing in data and count cells that
express each combination and sort combinations by the most to least frequent}
\usage{
markerComboCounts(dataFiles, markerFile, outFile = NULL)
}
\arguments{
\item{dataFiles}{vector of *.rda files containing Halo data}

\item{markerFile}{file containing a single identity marker on each line}

\item{outFile}{name of *.xlsx file to write final tables; default=NULL}
}
\value{
list of tables, one for all samples and one for each sample
}
\description{
Each row contains cell count, percentage and cumulative percentage of all cells,
and a column for each marker, with an "X" indicating that marker is part of the
combination in that row
}
