% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_combo_table.R
\name{markerComboWorkbook}
\alias{markerComboWorkbook}
\title{Create XLSX workbook and style all sheets for marker combo tables}
\usage{
markerComboWorkbook(allTbls, cellTypes)
}
\arguments{
\item{allTbls}{list of tables, one for each sheet, each containing
the following columns: Count, PCT, CUM.PCT, Interpretation, 
and a column for each marker being counted}
}
\value{
a completely styled XLSX workbook
}
\description{
For each marker combo table, create a XLSX sheet and style
according to original specifications (TO DO: add options?)
}
