% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_combo_table.R
\name{markerComboXLSX}
\alias{markerComboXLSX}
\title{Generate XLSX file of all marker combinations represented in
given halo data files}
\usage{
markerComboXLSX(dataFiles, markerFile, cellTypesFile, unreasonableCombosFile,
  outFile)
}
\arguments{
\item{dataFiles}{vector of *.rda files containing Halo data}

\item{markerFile}{file containing a single identity marker on each line}

\item{cellTypesFile}{a *.xlsx file containing sheets for Simple and Complex
marker combinations that identify certain cell types (format
described in docs)}

\item{unreasonableCombosFile}{a *.xlsx file containing a matrix that indicated invalid or
"unknown" pairwise combinations of markers in markerFile (format
described in docs)}

\item{outFile}{name of *.xlsx file to write final tables}
}
\description{
XLSX file generated will contain a detailed count sheet and a summary sheet for ALL data, 
incluing all samples, and for each sample. Each sheet includes marker count, percentage 
and cumulative percentage of all cells that contain each marker combination, and interpretation
or cell type of each marker combination
}
