% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{markerStringToPredicate}
\alias{markerStringToPredicate}
\title{Generate string for selecting markers based on positive
or negative value in data}
\usage{
markerStringToPredicate(ss)
}
\description{
Generate string for selecting markers based on positive
or negative value in data
}
