% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.R
\name{markExclusions}
\alias{markExclusions}
\title{Given a tibble of object analysis data, add a column EXCLUDE to indicate
which cells should be excluded from further analysis}
\usage{
markExclusions(samp, dat, drift, fovAnn, cellDiveId, haloAnn = NULL,
  borderPad = 20, driftThreshold = 0.1, printPlots = FALSE,
  debugDir = getwd(), boundaryColors = NULL,
  boundaryReassignmentFile = NULL)
}
\arguments{
\item{samp}{sample name as it appears in both SampleAnnotations.xlsx and data}

\item{dat}{tibble containing Halo object analysis data, to which EXCLUDE column
will be added}

\item{drift}{tibble of drift/loss summary}

\item{fovAnn}{FOV annotations for one sample}

\item{cellDiveId}{CELL_DIVE_ID mapping sample name to fov annotations}

\item{borderPad}{number in pixels indicating the minimum distance between a cell and the FOV
border in order for that cell NOT to be excluded}

\item{driftThreshold}{maximum pixel percent determined to have drifted in order to NOT be excluded}
}
\value{
tibble matching {dat} exactly, with an extra column, EXCLUDE, added to indicate which
        cells should not be analyzed
}
\description{
Given a tibble of object analysis data, add a column EXCLUDE to indicate
which cells should be excluded from further analysis, according to various upstream
factors including drift/loss, border padding, and other technical reasons
determined by investigators
}
