% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halo.R
\name{medianRow}
\alias{medianRow}
\title{Create median row}
\usage{
medianRow(dat, samp)
}
\arguments{
\item{dat}{table of counts where first three columns are Sample, 
FOV, SLICE, and the remaining columns are marker counts or fractions}

\item{samp}{sample name}
}
\value{
a row where Sample is the unique sample in input table, FOV and SLICE are NA,
          and the remaining values are medians of each column
}
\description{
Given a tibble with the first three columns Sample, FOV, SLICE, 
caclulate medians of all remaining columns and return row to be added to tibble
}
