% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{opponentprocess}
\alias{opponentprocess}
\title{opponentprocess()}
\usage{
opponentprocess(
  ii = 10000,
  sim_length = 10000,
  addl = 10000,
  plot_utility = FALSE,
  plot_op = FALSE,
  join_plots = TRUE,
  k_Dose = 1,
  k_apk = 0.01,
  k_bpk = 0.01,
  k_apd = 1,
  k_bpd = 2,
  k_H = 1,
  lambda_a = 1,
  gamma_a = 0.5,
  lambda_b = 2,
  gamma_b = 0.7,
  infuse = 1,
  plot_frequencies = c(0.002, 0.006),
  colorscheme = 1,
  verbose = FALSE
)
}
\arguments{
\item{ii}{Dosing interval.}

\item{sim_length}{Time length of PKPD simulation, in minutes.}

\item{addl}{Number of additional doses to deliver.}

\item{plot_utility}{Whether to calculate the graphs for biophase or not.}

\item{plot_op}{Whether to plot the graph of opponent processes (Hedonic state, H compartment)}

\item{join_plots}{Whether to join plots as subplots or plot them separately.}

\item{k_Dose}{Clearance rate for compartments.}

\item{k_apk}{a-process pharmacokinetic clearance rate.}

\item{k_bpk}{b-process pharmacokinetic clearance rate.}

\item{k_apd}{a-process pharmacodynamic clearance rate.}

\item{k_bpd}{b-process pharmacodynamic clearance rate.}

\item{k_H}{Clearance rate for hedonic pharmacodynamic compartment.}

\item{lambda_a}{Pharmacodynamic constant for a-process magnitude, based on Kahneman/Tversky's utility function.}

\item{gamma_a}{Pharmacodynamic constant for a-process curvature, based on Kahneman/Tversky's utility function.}

\item{lambda_b}{Pharmacodynamic constant for b-process magnitude, based on Kahneman/Tversky's utility function.}

\item{gamma_b}{Pharmacodynamic constant for b-process curvature, based on Kahneman/Tversky's utility function.}

\item{infuse}{Infusion duration.}

\item{plot_frequencies}{# If the dose frequency (or frequencies) are in plot_frequencies, then create plots for those frequencies. Should be a vector of length 2 - will always plot}

\item{colorscheme}{# Sets the color scheme; integer between 1 and 5.}

\item{verbose}{# Print output to console.}
}
\value{
A list containing the integral of hedonic outcomes, dose frequency, and optional plots.
}
\description{
opponentprocess() simulates the opponent process model. It can be run by itself for diagnostic purposes (for example, calculating the area under the hedonic curve, or plotting either the opponent processes or utility function), but in general use this should be called through bode_plot().
}
\examples{
opponentprocess() # Default parameters
opponentprocess(ii=10, sim_length=4000, addl=10000, plot_utility=FALSE, join_plots=TRUE, k_Dose=1, k_apk=0.01, k_bpk=0.01, k_apd=1, k_bpd=0.01, k_H=1, lambda_a=1, gamma_a=0.5, lambda_b=1, gamma_b=0.7, infuse=1, plot_frequencies=c(0.002), verbose=FALSE) # Default parameters
opponentprocess(plot_utility=TRUE) # Plots utility function
opponentprocess(plot_op=TRUE) # Plots hedonic compartment values (opponent processes)

}
