% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parseManifestList}
\alias{parseManifestList}
\title{Convert manifest value string to list}
\usage{
parseManifestList(listString)
}
\arguments{
\item{listString}{string representing the list to be returned}
}
\value{
list form of the string given
}
\description{
Lists can be written out in the manifest in the following form:
  Name1:val1,Name2:val1+val2,Name3:val1+val2+val3
This function parses this string to give the following list:
  list(Name1=val1, Name2=c(val1,val2), Name3=c(val1,val2,val3))
}
