% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_marker_percentages}
\alias{plot_marker_percentages}
\title{Make a grid of pie charts, one for each given marker showing
the percentages of all the marker combinations in which that 
positive marker is included}
\usage{
plot_marker_percentages(countsTable, markerNames, type = "bar",
  other_threshold = 0.05, exclude_sample_fov = NULL, pdfFile = NULL,
  v = TRUE, logFile = NULL, debug = FALSE, custom_colors = FALSE)
}
\arguments{
\item{countsTable}{tibble of counts generated by countMarkers()}

\item{markerNames}{vector of positive markers, for each of which a pie will be made}

\item{type}{type of chart ["bar"|"pie"]; default="bar"}

\item{other_threshold}{all markers making up less that this percentage will be put togther in 
a marker combination named 'other'; default=0.05}

\item{exclude_sample_fov}{a string of Sample FOVs to exclude from analysis in the form: 
'Sample1:2+4+3,Sample2:1,Sample3:2+4'}

\item{pdfFile}{output PDF file}

\item{v}{verbose - when set to TRUE, messages
will be printed to screen as well as to file;
DEFAULT = TRUE}

\item{logFile}{file to write log messages to; DEFAULT=NULL}

\item{debug}{print debug messages; DEFAULT=FALSE}

\item{custom_colors}{use color scheme defined in get_marker_colors(); TO DO: CHANGE THIS}
}
\description{
Make a grid of pie charts, one for each given marker showing
the percentages of all the marker combinations in which that 
positive marker is included
}
