% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{plotDensity}
\alias{plotDensity}
\title{Plot density of each FOV for one sample}
\usage{
plotDensity(dataFiles, annotationsDirs, cellTypeFiles, cellTypeName, fovBB, pad,
  pdfFile, logPlot = F, funcMarker = NULL, funcPosColor = "grey50",
  funcNegColor = "grey35", sampleColor = "#f16913",
  sampleColorDark = "#80380A", sortByMarker = "CD3", sampleOrder = NULL,
  exclude_sample_fov = NULL, exclude_sample_marker = NULL,
  writeCSVfiles = TRUE, maxG = 5, outDir = getwd(), byBand = FALSE,
  bandWidth = 10, maxDistanceFromInterface = 200, densityFiles = NULL,
  boxPlots = TRUE, singlePanelPlots = TRUE, includeSortByMarker = TRUE,
  consistentYScaling = TRUE, fileSetup = "single")
}
\arguments{
\item{dataFiles}{vector of *.rda files containing Halo data for a single sample}

\item{annotationsDirs}{vector of directories containing *.annotations files, XML files of boundary
annotations from Halo; order must correspond to order of dataFiles}

\item{cellTypeFiles}{a vector of text files with each line containing a cell type marker; may be
marker combinations delimited by commas}

\item{cellTypeName}{a character string naming the group of cell types (e.g., T Cells)}

\item{fovBB}{a list of four elements: X0,X1,Y0,Y1, representing the min/max coords
for FOVs}

\item{pad}{amount to trim from each FOV}

\item{pdfFile}{output PDF file}

\item{logPlot}{logical; when set to TRUE, plot density per mm^2; default=FALSE}

\item{funcMarker}{character string of functional marker to plot on top of every other marker}

\item{sampleColor}{character string indicating HTML color to represent this particular sample}

\item{sampleColorDark}{a slightly darker shade of sampleColor, to be used for density of funcMarker+}

\item{sortByMarker}{sort FOVs by density of this marker (cell type)in ascending order;
default=CD3}

\item{exclude_sample_fov}{list where names are sample names and values are vectors of FOV to be
removed from data before any calculations}

\item{exclude_sample_marker}{list where names are sample names and values are vectors of markers to be 
should be removed from data before any calculations}

\item{writeCSVfiles}{logical; write CSV files, one of density values and one of area values; 
default=TRUE}

\item{maxG}{maximum factor of 10 to use for generating random points on grid [ REWORD??? ]}

\item{outDir}{if writeCSVfiles=T, directory to which these files will be written#' @return nothing}

\item{byBand}{logical indicating whether to break down density by distance from tumor interface; 
default=FALSE}

\item{bandWidth}{width of each band around tumor interface, to be used when byBand is TRUE; 
default=10}

\item{maxDistanceFromInterface}{include only those cells within this distance from tumor interface}

\item{boxPlots}{logical indicating whether to generate a box plot for each cell type
comparing mean densities between samples; default=TRUE}

\item{singlePanelPlots}{logical; when set to FALSE, the plot for each marker will be printed separately; 
by default, there will be one page of plots for all cell types in each cellTypeFile}

\item{consistentYScaling}{logical; when onePanel=FALSE, set this parameter to FALSE in order to scale separate 
plots individually; default=TRUE}

\item{fileSetup}{['single'|'per-marker'|'per-cellTypeFile']; when set to 'single', ALL plots will be in
a single PDF file; must provide pdfFile}
}
\description{
Plot density of each FOV for one sample
}
