% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{plotDensityByBand}
\alias{plotDensityByBand}
\title{Plot density broken down by distance from tumor interface}
\usage{
plotDensityByBand(density, sampleName, markers, pdfFile = NULL)
}
\arguments{
\item{density}{tibble with columns for Density, Band and CellType}

\item{sampleName}{name of sample from which the FOVs came from}

\item{markers}{vector of markers/marker combos}

\item{pdfFile}{pdfFile to which plot will be printed}
}
\description{
Given a vector of markers and a tibble containing columns for Density and 
Band, generate a bar plot for each FOV, with one bar per Band, colored by 
marker
}
