% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.R
\name{plotExclusions}
\alias{plotExclusions}
\title{Plot cell locations of exclusions}
\usage{
plotExclusions(dd, haloAnnotations = NULL, iFiles = NULL, aFiles = NULL,
  epFiles = NULL, gFiles = NULL, outDir = getwd(),
  boundaryColors = NULL, boundaryReassignmentFile = NULL)
}
\arguments{
\item{dd}{data for one sample with EXCLUDE column}

\item{haloAnnotations}{halo annotations in list format, for one sample and one or multiple FOV}

\item{iFiles}{if not providing haloAnnotations, a vector of interface XML files}

\item{aFiles}{if not providing haloAnnotations, a vector of exclusion XML files}

\item{epFiles}{if not providing haloAnnotations, a vector of epidermis XML files}

\item{gFiles}{if not providing haloAnnotations, a vector of glass XML files}

\item{outDir}{directory where plots should be saved}

\item{boundaryColors}{list of boundary colors as they appear in halo annotation XML files
where the values are the associated boundary code (e.g., 'epiB', 'glsB', etc.)}

\item{boundaryReassignmentFile}{comma-separated file indicating exceptions to the color assignments
in boundaryColors}
}
\description{
Plot cell locations of exclusions
}
