% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotMarkerPercentages}
\alias{plotMarkerPercentages}
\title{Make a grid of pie charts, one for each given marker showing
the percentages of all the marker combinations in which that 
positive marker is included}
\usage{
plotMarkerPercentages(countsTable, markerNames, type = "bar",
  other_threshold = 0.05, exclude_sample_fov = NULL, pdfFile = NULL,
  custom_colors = FALSE)
}
\arguments{
\item{countsTable}{tibble of counts generated by countMarkers()}

\item{markerNames}{vector of positive markers, for each of which a pie will be made}

\item{type}{type of chart ["bar"|"pie"]; default="bar"}

\item{other_threshold}{all markers making up less that this percentage will be put togther in 
a marker combination named 'other'; default=0.05}

\item{exclude_sample_fov}{a string of Sample FOVs to exclude from analysis in the form: 
'Sample1:2+4+3,Sample2:1,Sample3:2+4'}

\item{pdfFile}{output PDF file}

\item{custom_colors}{use color scheme defined in getMarkerColors(); TO DO: CHANGE THIS}
}
\description{
Make a grid of pie charts, one for each given marker showing
the percentages of all the marker combinations in which that 
positive marker is included
}
