% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{plotTotalDensity}
\alias{plotTotalDensity}
\title{Plot density for each marker, broken down by FOV}
\usage{
plotTotalDensity(markers, markerTable, funcMarker = NULL,
  funcPosColor = "grey35", funcNegColor = "grey50",
  sampleColor = "#f16913", sampleColorDark = "#80380A", spotOrder = NULL,
  sampleOrder = NULL, onePanel = TRUE, pdfFile = NULL)
}
\arguments{
\item{markers}{vector of markers/marker combinations to be plotted; there will
be one plot per marker}

\item{markerTable}{tibble containing columns for Sample, SPOT, and a column of densities
for each marker}

\item{funcMarker}{a functional marker that is included in the marker combos in markerTable;
for each marker, there should be a combo that includes the functional marker
(indicating that cell type is positive for this marker) and one that does
not (indicating that the cell type may or may not be positive for this marker);
default=NULL;}

\item{sampleColor}{all bars will be colored gray except for the Median bar, which will be
this color; default=#f16913}

\item{sampleColorDark}{a darker version of sampleColor (or really any color of your choosing)
that will represent the median of cells positive for the funcMarker;
default=#80380A}

\item{spotOrder}{a vector of FOV numbers indicating the order in which FOVs should be in 
the plot(s); default=NULL}

\item{sampleOrder}{a vector indicating the order in which samples should appear}

\item{onePanel}{if plotting multiple markers, logical indicating all plots should be on one
panel; default=TRUE}

\item{pdfFile}{if given, plot(s) will be written to this file; default=NULL}

\item{logPlot}{logical indicating that density should be plotted on log scale; default=FALSE}
}
\description{
Bar plot(s) showing density of a marker in each FOV of one sample
}
