% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{plotTotalDensity__OLD}
\alias{plotTotalDensity__OLD}
\title{Plot total density of all bands around a tumor interface for one FOV}
\usage{
plotTotalDensity__OLD(markers, markerTable, ymax, funcMarker = NULL,
  sampleColor = "#f16913", sampleColorDark = "#80380A", spotOrder = NULL,
  logPlot = FALSE, pdfFile = NULL)
}
\arguments{
\item{markers}{vector of markers/marker combos}

\item{markerTable}{tibble containing columns for [ FILL THIS IN LATER ]}

\item{ymax}{maximum y value; by default, this value is dictated by the data}

\item{funcMarker}{functional marker to plot on top of every other marker [ FIX THIS ]}

\item{sampleColor}{character string indicating HTML color to represent this particular sample}

\item{sampleColorDark}{a slightly darker shade of sampleColor, to be used for density of funcMarker+}

\item{spotOrder}{vector of FOVs in the order in which they should appear on the plot}

\item{logPlot}{plot density on log scale}

\item{pdfFile}{PDF file to which plot should be saved; default=NULL}
}
\description{
Plot total density of all bands around a tumor interface for one FOV
}
