% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTotalFOVMarkerDensity}
\alias{plotTotalFOVMarkerDensity}
\title{Plot and print absolute density or density percentages of cell types in entire FOVs}
\usage{
plotTotalFOVMarkerDensity(mDen, densityMarkers, clrs, plotTitle = "",
  cellTypeLabels = NULL, sampleOrder = NULL, yMax = NULL,
  separateLegend = FALSE, printLegend = TRUE, legendOnly = FALSE,
  xCol = "Sample", yCol = "Total", pct = FALSE, groupBy = NULL,
  yAxisTitle = "Density (counts/mm^2)", xAxisTitle = "Sample",
  annotationCols = NULL)
}
\arguments{
\item{mDen}{tibble containing density data}

\item{densityMarkers}{a vector of markers for which density is to be plotted}

\item{clrs}{a list of colors, named by either by CellType from mDen tibble, 
or cellTypeLabels if specified}

\item{plotTitle}{title of plot}

\item{cellTypeLabels}{a list of labels for each marker being plotted, if different
than CellType in mDen tibble; default: NULL}

\item{sampleOrder}{order in which samples should appear on the plot; default=NULL}

\item{yMax}{if specified, y axis will span from zero to this number; default=NULL}

\item{separateLegend}{logical; when TRUE, legend will be separated from plot. if printLegend is TRUE, 
will be printed first, before plot. if printLegend is FALSE, will not be printed at all; default=FALSE}

\item{printLegend}{logical; when separateLegend is TRUE, if printLegend is FALSE, no legend will be printed at all; default=TRUE}

\item{legendOnly}{logical; when separateLegend is TRUE, if legendOnly is TRUE, only legend will be printed, no plot}

\item{yCol}{character string specifying which of the columns in mDen is to be used for y values; default="Density"}

\item{pct}{logical; values being plotted are percentages; default=FALSE}

\item{yAxisTitle}{character string to be used as title of the y axis}

\item{xAxisTitle}{character string to be used as title of the x axis}

\item{annotationCols}{vector of column names to use for annotation}
}
\value{
plot
}
\description{
Plot and print absolute density or density percentages of cell types in entire FOVs
}
