% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{printDensityPlots}
\alias{printDensityPlots}
\title{Print density plots}
\usage{
printDensityPlots(den, config, yScaleConsistency = "byPopulation",
  absoluteDensity = TRUE, densityPercentages = TRUE, byFOV = TRUE,
  summarize = TRUE)
}
\arguments{
\item{den}{tibble containing densities for all markers to be plotted}

\item{config}{configuration of plots (already parsed)}

\item{yScaleConsistency}{level on which y-scales should be the same; choices: ["byPopulation"|"all"|"markerSet"]}

\item{absoluteDensity}{logical; print plots of absolute density values; default=TRUE}

\item{densityPercentages}{logical; print plots of density percentage values; default=TRUE}

\item{byFOV}{logical; print plots of each FOV (all FOV on one panel); default=TRUE}

\item{summarize}{logical; print summary plots, showing density of ALL FOV together; default=TRUE}

\item{stacked}{logical; for plots with all exclusive markers, print stacked bar plots; default=TRUE}
}
\value{
nothing
}
\description{
Print density plots
}
