% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{printTotalDensityPlots}
\alias{printTotalDensityPlots}
\title{Print bar charts showing marker densities for total FOVs}
\usage{
printTotalDensityPlots(den, areas, markerConfig,
  yScaleConsistency = "population", absoluteDensity = TRUE,
  densityPercentage = TRUE, byFOV = TRUE, summarize = TRUE,
  stacked = TRUE, sampleOrder = NULL, separateLegend = TRUE,
  outDir = NULL)
}
\value{
nothing
}
\description{
For each population in each marker set, plot total FOV densities of specified marker combinations.
One plot will be printed for a single population in a single sample, comparing densities of functional markers (?)
}
