% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{printTotalFOVDensityPlots}
\alias{printTotalFOVDensityPlots}
\title{Print plots of total FOV marker density plots}
\usage{
printTotalFOVDensityPlots(den, fovAreas, config,
  yScaleConsistency = "population", indivPopulations = TRUE,
  absoluteDensity = TRUE, densityPercentage = TRUE, summarize = TRUE,
  stacked = TRUE, forceStack = FALSE, sampleOrder = NULL, outDir = NULL,
  annotationCols = NULL)
}
\arguments{
\item{den}{tibble containing densities for all markers to be plotted}

\item{fovAreas}{tibble containing area data for all Samples and FOVs}

\item{config}{tibble containing marker configuration (returned from getMarkerConfig())}

\item{yScaleConsistency}{level on which y-scales should be the same; choices: ["byPopulation"|"all"|"markerSet"]}

\item{indivPopulations}{logical; print a plot for each individual population; default=TRUE}

\item{absoluteDensity}{logical; print plots of absolute density values; default=TRUE}

\item{densityPercentage}{logical; print plots of density percentage values; default=TRUE}

\item{summarize}{logical; print summary plots, showing density of ALL FOV together; default=TRUE}

\item{stacked}{logical; for plots with all exclusive markers, print stacked bar plots; default=TRUE}

\item{forceStack}{force stacking of cell types even if they appear to be unstackable}

\item{sampleOrder}{vector of samples in order that they should appear on plot}

\item{outDir}{directory where all plots will be printed, organized by marker set}

\item{annotationCols}{vector of column names to use for annotation}
}
\value{
nothing
}
\description{
Print plots of total FOV marker density plots
}
