% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{projectFileName}
\alias{projectFileName}
\title{Generate name of output file for marker stats based on 
input file names}
\usage{
projectFileName(markerFile, dataFiles, pad, type)
}
\arguments{
\item{markerFile}{File containing list of marker names}

\item{dataFiles}{vector of data file(s)}

\item{pad}{amount that will be trimmed from FOV}

\item{type}{file type ("rda","txt","xlsx"); essentially
the file extension}
}
\value{
file name to be used for saving counts
}
\description{
Use name of marker file and either a single data file name
or the directory of data files to generate output file name
}
