% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{projectParams}
\alias{projectParams}
\title{Build list of project parameters and values by reading
project manifest}
\usage{
projectParams(file)
}
\arguments{
\item{file}{manifest file to be read}
}
\value{
a list of all project parameters
}
\description{
Given a tab-delimited file of key-value pairs, read
them into list. File may contain comments both on top and in columns
beyond the second, but will be ignored here
}
\details{
TO DO: Add to description list of required and optional keys in file
}
