% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{remove_exclusions}
\alias{remove_exclusions}
\title{Remove from counts tibble any FOV to be excluded}
\usage{
remove_exclusions(dat, exclusions, v = TRUE, debug = FALSE)
}
\arguments{
\item{dat}{counts tibble from countMarkers()}

\item{exclusions}{a string of exlusions in the form: Sample1:3+5+9,Sample2:1+16+22}

\item{v}{verbose - when set to TRUE, messages
will be printed to screen as well as to file;
DEFAULT = TRUE}

\item{debug}{print debug messages; default=FALSE}
}
\description{
Filter data to exclude specific FOV for specific samples
}
