% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{removeExcludedPoints}
\alias{removeExcludedPoints}
\title{Remove points from data set that fall inside exclusion boundaries}
\usage{
removeExcludedPoints(ds, aFile = NULL, excB = NULL)
}
\arguments{
\item{ds}{tibble containing at minimum, Sample, SPOT, UUID, X, Y, Marker, Value}

\item{aFile}{Halo boundaries annotation file in XML format; if NULL, must provide
the list of excluded region tables as returned by cleanBoundaries()}

\item{excB}{a list of excluded region tables as returned by cleanBoundaries;
if NULL, must provide the original Halo annotations file in XML format}
}
\value{
tibble filtered to remove points that fall inside exclusion boundaries
}
\description{
Read Halo boundaries annotations file and remove from data set any cells
that fall inside those regions
}
