% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sampleInfiltrationPlotsSingleFile}
\alias{sampleInfiltrationPlotsSingleFile}
\title{Organize and print multiple sample summaries on a page or two}
\usage{
sampleInfiltrationPlotsSingleFile(den, allPlots, numPtsPerPage = 4,
  ncol = NULL, nrow = NULL)
}
\arguments{
\item{den}{density table including Sample_name, Patient, and Response in addition to standard density data
NOTE: currently Sample_name MUST be in the form "Pt[X]_[ResponseX]", e.g., "Pt1_UT1" or "Pt1_PR"}

\item{allPlots}{list of plots organized by markerSet, population, sample, [summary], [absolute|percentage]}

\item{numPtsPerPage}{number of cases to plot on each page}
}
\value{
nothing
}
\description{
For each marker set in list of plots organized by markerSet, population, then sample,
gather all sample plots for a single population, organizing with each row including
a single patient and each column including a treatment
}
