% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stackable}
\alias{stackable}
\title{Determine whether a set of markers is 'stackable' in a bar chart}
\usage{
stackable(markers)
}
\arguments{
\item{markers}{vector of marker combinations to be tested for stackability}
}
\value{
subset of markers including only the ones that ARE stackable
}
\description{
Extract from a set of markers only those combinations that are mutually exlusive and therefore 
'stackable' in a bar chart
}
