% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{summarizeByCellTypeDefinition}
\alias{summarizeByCellTypeDefinition}
\title{Summarize positive/negative marker matrix by cell type definition}
\usage{
summarizeByCellTypeDefinition(assignments, bandArea, config, plotConfigFile)
}
\arguments{
\item{assignments}{table of 0s and 1s, where columns are markers (may be pos or neg)
and each row is a unique cell; zero indicates cell is NOT positive for that marker
(note, if marker name is negative a zero means the opposite) and one 
indicates positivity}

\item{bandArea}{table indicating calculated area for each sample, fov and band}

\item{config}{parsed marker config of info for each cell type definition including marker 
combinations (result of getMarkerConfig())}

\item{plotConfigFile}{full path to file containing plot options including colors for cell types}
}
\value{
list of two items: den=density table, config=updated config to represent summary info
}
\description{
Get cell counts and densities by cell definition rather than marker combination
}
