% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{summarizeFOVDataByCellTypeDefinition}
\alias{summarizeFOVDataByCellTypeDefinition}
\title{Summarize density data by cell type definition}
\usage{
summarizeFOVDataByCellTypeDefinition(den, config, areas, plotConfigFile)
}
\arguments{
\item{den}{density table}

\item{config}{parsed marker config of info for each cell type definition including marker 
combinations (result of getMarkerConfig())}

\item{areas}{fovAreas}

\item{plotConfigFile}{full path to file containing plot options including colors for cell types}
}
\value{
list of two items: den=density table, config=updated config to represent summary info
}
\description{
Summarize density data by cell type definition
}
