% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halo.R
\name{trimImage}
\alias{trimImage}
\title{Remove cells near edge of FOV from dataset}
\usage{
trimImage(dat, pad = 0)
}
\arguments{
\item{dat}{dataset (tibble)}

\item{pad}{amount to trim around edges (pixels, double; will be converted
to um)}
}
\value{
tibble with cells that fall outside padding filtered out
}
\description{
Given a tibble and an amount to trim around edges, remove cells
that fall outside the inner border
}
