% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validateConfig}
\alias{validateConfig}
\title{Validate study configuration}
\usage{
validateConfig(config)
}
\arguments{
\item{config}{study configuration in list format}
}
\value{
updated configuration in list format
}
\description{
Validate study configuration and set any available defaults
}
