% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validateExclusions}
\alias{validateExclusions}
\title{Validate data exclusions}
\usage{
validateExclusions(dat)
}
\arguments{
\item{dat}{data frame including EXCLUDE column}
}
\value{
nothing
}
\description{
Given a data frame with an EXCLUDE column, count number and fraction
of all possible types of exclusions
}
