% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halo.R
\name{validateMarkers}
\alias{validateMarkers}
\title{Modify data to sync to markers in marker file}
\usage{
validateMarkers(dat, m2, v = FALSE, logFile = NULL)
}
\arguments{
\item{dat}{tibble containing marker data}

\item{m2}{list of individual markers in marker file}

\item{v}{verbose - when set to TRUE, messages
will be printed to screen as well as to file;
DEFAULT = TRUE}

\item{logFile}{file to write log messages to; DEFAULT=NULL}
}
\value{
a tibble with extra markers filtered out
}
\description{
Filter out any markers in data that are not in marker file; Add to data
markers in marker file but not  
Print a warning with a list of any markers that are in data but not
marker file OR that are in marker file but not data
}
