% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.R
\name{writeExclusionReasons}
\alias{writeExclusionReasons}
\title{Add exclusion reasons to a vector of existing exclusion reasons}
\usage{
writeExclusionReasons(exclCol, idxs, reason)
}
\arguments{
\item{exclCol}{vector of existing exclusions (EXCLUDE column from object analysis data)}

\item{idxs}{vector of indices at which exclusion reason should be added}

\item{reason}{string to be added to exclude column indicating why a cell is to be excluded}
}
\value{
vector of exclusion reasons with new reason added
}
\description{
Given an exclusion reason and the indices of cells to be excluded, 
properly add the reason to vector of existing exclusions
}
