% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halo.R
\name{writeXLSXsheet}
\alias{writeXLSXsheet}
\title{Shortcut to write a XLSX file}
\usage{
writeXLSXsheet(dat, outFile, sheetName = NULL, append = TRUE)
}
\arguments{
\item{dat}{data to be written, either tibble, matrix or dataframe}

\item{outFile}{name of output file}

\item{sheetName}{name of Excel Worksheet}

\item{append}{logical indicating whether to append to the file or
overwrite; Default=TRUE}
}
\description{
Write a XLSX file with the following defaults:
   row.names = F
   append = TRUE
   sheetName = NULL
}
