% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halo.R
\name{addMissingMarkers}
\alias{addMissingMarkers}
\title{Add missing markers}
\usage{
addMissingMarkers(dat, m2)
}
\arguments{
\item{dat}{tibble containing marker data}

\item{m2}{list of individual markers in marker file}
}
\value{
a tibble with missing markers added, with "Positive" value
         set to NA
}
\description{
Add to data markers that are in marker file but are missing
from given data set
}
