% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{calculateFOVStats}
\alias{calculateFOVStats}
\title{Calculate total area and density of entire FOVs}
\usage{
calculateFOVStats(aFiles, data, markerSet, cellTypeName, pad, maxG,
  writeCSVfiles = TRUE, funcMarker = NULL, sortByMarker = NULL, outDir)
}
\arguments{
\item{aFiles}{a vector of Halo boundary annotation files, with 
'.annotations' extension in XML format}

\item{data}{Halo data for one sample loaded from a *.rda}

\item{markerSet}{a vector markers to calculate density for}

\item{cellTypeName}{a name to represent the cell types in marker file}

\item{pad}{amount to trim from FOV before calculating}

\item{maxG}{the maximum factor of 10 to use for generating random points for area calculation}

\item{writeCSVfiles}{logical indicating whether to write density and area tables to file;}

\item{funcMarker}{this marker will be added to all markers in markerSet in order to
compare cells that are +/- for this functional marker [ TO DO: REWORD THIS ]}

\item{outDir}{if writeCSVfiles=T, directory to which these files will be written}
}
\value{
a list containing two tables: 'density' and 'area'
}
\description{
Calculate total area and density of FOVs that do not contain 
tumors
}
