% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{calculateInterfaceStats}
\alias{calculateInterfaceStats}
\title{Calculate total area and density of tumor interface}
\usage{
calculateInterfaceStats(aFiles, dataFile, cellTypeFile, cellTypeName,
  fovBoundingBox, pad, haloInfiltrationDir, funcMarker = NULL,
  writeCSVfiles = TRUE, maxG = 5, outDir = NULL,
  interfaceBins = (-20:20) * 10, statsByBand = FALSE)
}
\arguments{
\item{aFiles}{a vector of Halo boundary annotation files, with 
'.annotations' extension in XML format}

\item{dataFile}{a *.rda file with Halo data for one sample}

\item{cellTypeFile}{file containing cell types to be calculated; each line
is one cell type and each cell type can be either a single
marker name or a comma-separated list of markers}

\item{cellTypeName}{a name to represent the cell types in marker file}

\item{fovBoundingBox}{a list with four elements: X0, X1, Y0, Y1, representing the minimum
and maximum coordinates of the FOV to be plotted}

\item{pad}{amount to trim from FOV before calculating}

\item{haloInfiltrationDir}{directory containing Halo Infiltration data (csv, plot files)}

\item{funcMarker}{this marker will be added to all markers in cellTypeFile in order to
compare cells that are +/- for this functional marker [ TO DO: REWORD THIS ]}

\item{writeCSVfiles}{logical indicating whether to write density and area tables to file; 
default=TRUE}

\item{maxG}{the maximum factor of 10 to use for generating random points for area calculation}

\item{outDir}{if writeCSVfiles=T, directory to which these files will be written}

\item{interfaceBins}{a vector of distances from tumor interface in which cells will be binned; 
default=(-20:20)*10}

\item{statsByBand}{logical; return stats broken down by interface bands (bins) as opposed to total
stats; default=FALSE}
}
\value{
a list containing two tables: 'density' and 'area'
}
\description{
Calculate area and density of cells within a given distance to a tumor boundary
}
