% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_counts.R
\name{cleanData}
\alias{cleanData}
\title{Clean data table}
\usage{
cleanData(dat)
}
\arguments{
\item{dat}{tibble containing marker data}
}
\value{
modified tibble
}
\description{
Alter column names, ensure correct data types for certain columns
}
