% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{g_legend}
\alias{g_legend}
\title{Extract a legend from a gplot}
\usage{
g_legend(plt)
}
\arguments{
\item{plt}{a plot generated with ggplot}
}
\description{
Extract a legend from a plot generated with ggplot
}
