% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{get_HaloDotAPI_token}
\alias{get_HaloDotAPI_token}
\alias{set_HaloDotAPI_token}
\title{Halo Infinite API Token Management}
\usage{
get_HaloDotAPI_token()

set_HaloDotAPI_token(token = NULL)
}
\arguments{
\item{token}{HaloDotAPI Token.}
}
\value{
\code{get_HaloDotAPI_token()} returns the API token if it is
found as an environment variable.
}
\description{
Get or set your HaloDotAPI access token as an environment variable.
}
