% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_spatial.R
\name{getBandAssignments}
\alias{getBandAssignments}
\title{Get counts of cells that fall within given distances of
tumor boundaries for a single FOV}
\usage{
getBandAssignments(ds, bbData, allBoundaries, interfaceBins = (-20:20) * 10)
}
\arguments{
\item{ds}{a tibble containing data for one FOV, including columns
for X, Y, and one for each marker}

\item{bbData}{a list containing X0,X1,Y0,Y1 representing the boundary
of the trimmed FOV}

\item{allBoundaries}{a list returned from readHaloAnnotations}

\item{interfaceBins}{vector of distances (integers) that define each band; default = (-20:20)*10}
}
\description{
Given a set of interface bands, or set distances from a tumor 
boundary, count the numbers of cells that fall within those bands
}
