% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_meta.R
\name{getTemplateStudyConfig}
\alias{getTemplateStudyConfig}
\title{Create a study config template}
\usage{
getTemplateStudyConfig(study_dir = getwd(),
  set_default_directory_struct = TRUE, set_default_file_names = TRUE,
  study_name = NULL, writeYAML = TRUE)
}
\arguments{
\item{study_dir}{root study directory, where CURRENT study dir will be created}

\item{set_default_directory_struct}{logical; if TRUE, default study directory and subdirectories will be created}

\item{study_name}{name of study}

\item{writeYAML}{logical; write study config to YAML file}
}
\description{
Create a study config template, returning config in list format and optionally writing
config to yaml file. User may choose to set all parameters to NULL or to set any available
default values.
}
