% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halo-package.R
\docType{package}
\name{halo-package}
\alias{halo}
\alias{halo-package}
\title{halo: High-Level Interface to Halo Infinite API}
\description{
Request data from the Halo Infinite API.
}
\author{
\strong{Maintainer}: Matthew T. Warkentin \email{warkentin@lunenfeld.ca} (\href{https://orcid.org/0000-0001-8730-3511}{ORCID})

}
\keyword{internal}
