% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appearance.R
\name{halo_appearance}
\alias{halo_appearance}
\title{Halo Infinite Player Appearance}
\usage{
halo_appearance(
  gamertag,
  version = get_HaloDotAPI_version(),
  token = get_HaloDotAPI_token(),
  .progress = TRUE
)
}
\arguments{
\item{gamertag}{Character vector of gamertags.}

\item{version}{HaloDotAPI Version.}

\item{token}{HaloDotAPI Token.}

\item{.progress}{Logical. Should a progress bar be displayed? Passed on to
\code{furrr::future_map()}.}
}
\value{
A \code{list} with class \code{appearance}.
}
\description{
Returns targeted player's appearance on Halo Infinite.
}
