% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\name{halo_articles}
\alias{halo_articles}
\title{Halo Infinite Articles}
\usage{
halo_articles(
  language = "en-US",
  version = "0.3.6",
  token = get_HaloDotAPI_token()
)
}
\arguments{
\item{language}{Desired language.}

\item{version}{HaloDotAPI Version.}

\item{token}{HaloDotAPI Token.}
}
\value{
A \code{list} with class \code{articles}.
}
\description{
Returns available articles on Halo Infinite.
}
