% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_record.R
\name{halo_campaign}
\alias{halo_campaign}
\alias{halo_multiplayer}
\title{Halo Infinite Service Record}
\usage{
halo_campaign(gamertag, version = "0.3.6", token = get_HaloDotAPI_token())

halo_multiplayer(
  gamertag,
  filter = NULL,
  version = "0.3.6",
  token = get_HaloDotAPI_token()
)
}
\arguments{
\item{gamertag}{Player's Gamertag.}

\item{version}{HaloDotAPI Version.}

\item{token}{HaloDotAPI Token.}

\item{filter}{Filter.}
}
\value{
A \code{list} with either class \code{stats_service_record_campaign} or
\code{stats_service_record_multiplayer}.
}
\description{
Returns targeted player's service record on Halo Infinite.
}
