% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csrs.R
\name{halo_CSRS}
\alias{halo_CSRS}
\title{Halo Infinite CSRS}
\usage{
halo_CSRS(
  gamertag,
  season,
  version = get_HaloDotAPI_version(),
  token = get_HaloDotAPI_token()
)
}
\arguments{
\item{gamertag}{Player's Gamertag.}

\item{season}{Targeted season.}

\item{version}{HaloDotAPI Version.}

\item{token}{HaloDotAPI Token.}
}
\value{
A \code{list} with class \code{stats_csrs}.
}
\description{
Returns targeted player's CSRS on Halo Infinite.
}
