% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{halo_matches_list}
\alias{halo_matches_list}
\title{Halo Infinite Player Matches}
\usage{
halo_matches_list(
  gamertag,
  count = NULL,
  offset = NULL,
  mode = NULL,
  version = get_HaloDotAPI_version(),
  token = get_HaloDotAPI_token(),
  .progress = TRUE
)
}
\arguments{
\item{gamertag}{Character vector of gamertags.}

\item{count}{Number of items to retrieve.}

\item{offset}{Number of items to offset.}

\item{mode}{Targeted mode. One of \code{'matchmade'} or \code{'custom'}.}

\item{version}{HaloDotAPI Version.}

\item{token}{HaloDotAPI Token.}

\item{.progress}{Logical. Should a progress bar be displayed? Passed on to
\code{furrr::future_map()}.}
}
\value{
A \code{list} with class \code{stats_matches_list}.
}
\description{
Returns targeted player's matches history on Halo Infinite.
}
