% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaloDotAPI.R
\name{HaloDotAPI}
\alias{HaloDotAPI}
\title{Perform a Request to HaloDotAPI}
\usage{
HaloDotAPI(endpoint, ..., version = "0.3.6", token = get_HaloDotAPI_token())
}
\arguments{
\item{endpoint}{Endpoint.}

\item{...}{Query parameters passed on to the chosen \code{endpoint}.}

\item{version}{HaloDotAPI Version.}

\item{token}{HaloDotAPI Token.}
}
\value{
A \code{list} with a class that is specific to the chosen \code{endpoint}.
}
\description{
This function is the main tool for making requests to HaloDotAPI. Most
users will not use this function directly, but will instead use
functions that are specific to the available endpoints.
}
