% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatialFunction.R
\name{plotCellTypeLocations}
\alias{plotCellTypeLocations}
\title{Plot location of given cell types within one FOV}
\usage{
plotCellTypeLocations(dataFile, annotationsDir, cellTypesFile, pad = 30,
  pdfFile = NULL, verbose = FALSE, logFile = NULL, debug = FALSE)
}
\arguments{
\item{dataFile}{*.rda file containing ObjectAnalysisData}

\item{annotationsDir}{directory of *.annotations XML files from Halo for one sample}

\item{cellTypesFile}{text file containing a list of cell type markers, one on each line; 
each cell type marker may consist of an indivdual marker or a comma-}

\item{pad}{amount to trim from FOV; default=30}

\item{pdfFile}{name of output PDF file}

\item{verbose}{print progress messages; default=FALSE}

\item{logFile}{log file; default=NULL}

\item{debug}{print debug messages; default=FALSE}
}
\value{
nothing
}
\description{
Given a halo data *.rda file with XML annotations and a list of cell types, 
plot the X and Y coordinates of those cells showing locations relative to
each other and to tissue boundary
}
