% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCellTypeLocations}
\alias{plotCellTypeLocations}
\title{Plot location of given cell types within one FOV}
\usage{
plotCellTypeLocations(dataFile, annotationsDir, cellTypesFile, fovBB, plotBB,
  boundaryColors, cellTypeColors, pad = 30, pdfFile = NULL)
}
\arguments{
\item{dataFile}{*.rda file containing ObjectAnalysisData}

\item{annotationsDir}{directory of *.annotations XML files from Halo for one sample}

\item{cellTypesFile}{text file containing a list of cell type markers, one on each line; 
each cell type marker may consist of an indivdual marker or a comma-}

\item{fovBB}{a list of four elements: X0,X1,Y0,Y1, representing the min/max coords
for FOVs}

\item{plotBB}{a list of four elements: X0,X1,Y0,Y1, representing the min/max coords
for whole plots}

\item{boundaryColors}{a list of hexadecimal color codes for Tum, Exc, and Epi boundaries}

\item{cellTypeColors}{a list of hexadecimal color codes for each cell type in cellTypesFile}

\item{pad}{amount to trim from each FOV}

\item{pdfFile}{name of output PDF file}
}
\value{
nothing
}
\description{
Given a halo data *.rda file with XML annotations and a list of cell types, 
plot the X and Y coordinates of those cells showing locations relative to
each other and to tissue boundary
}
