% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{printInfiltrationDensityPlots}
\alias{printInfiltrationDensityPlots}
\title{Print density plots}
\usage{
printInfiltrationDensityPlots(den, bandWidth = NULL, config,
  yScaleConsistency = "population", absoluteDensity = TRUE,
  densityPercentage = TRUE, byFOV = TRUE, summarize = TRUE,
  stacked = TRUE, sampleOrder = NULL, separateLegend = TRUE,
  infiltrationAreas = NULL, bandAssignments = NULL, outDir = getwd())
}
\arguments{
\item{den}{tibble containing densities for all markers to be plotted}

\item{bandWidth}{if plotting by distance intervals from infiltration boundary, this is the size of 
each interval}

\item{config}{tibble containing marker configureation (returned from getMarkerConfig())}

\item{yScaleConsistency}{level on which y-scales should be the same; choices: ["byPopulation"|"all"|"markerSet"]}

\item{absoluteDensity}{logical; print plots of absolute density values; default=TRUE}

\item{densityPercentage}{logical; print plots of density percentage values; default=TRUE}

\item{byFOV}{logical; print plots of each FOV (all FOV on one panel); default=TRUE}

\item{summarize}{logical; print summary plots, showing density of ALL FOV together; default=TRUE}

\item{stacked}{logical; for plots with all exclusive markers, print stacked bar plots; default=TRUE}
}
\value{
nothing
}
\description{
Print density plots
}
