% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.R
\name{readHaloAnnotations}
\alias{readHaloAnnotations}
\title{Parse boundary annotations of a single FOV in a single XML file}
\usage{
readHaloAnnotations(annoteFile, annoteType)
}
\arguments{
\item{annoteFile}{XML file containing boundary annotations for a single FOV}
}
\description{
Parse XML file containing annotations for one FOV. Write tab-delimited file
of select information (see ?parseAnnotationXML for details)
}
