% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halo.R
\name{removeExtraMarkers}
\alias{removeExtraMarkers}
\title{Remove from data markers that are not in marker file}
\usage{
removeExtraMarkers(dat, m2)
}
\arguments{
\item{dat}{tibble containing marker data}

\item{m2}{list of individual markers in marker file}
}
\value{
a tibble with extra markers filtered out
}
\description{
Remove from data markers that are not in marker file
}
