% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halo.R
\name{validateSingleFOVFile}
\alias{validateSingleFOVFile}
\title{Make sure list of unique FOVs contains only one element}
\usage{
validateSingleFOVFile(uniqFOVs)
}
\arguments{
\item{uniqFOVs}{vector containing list of unique FOVs in dataset}
}
\description{
If the list is longer than one element, print error message and EXIT
}
